### 2017-11-04 #155 @adamhsparks
I was looking at the POWER website. The most, but not all, of the other POWER datasets have similar interfaces. It should be pretty simple add some of the Renewable Energy and Sustainable Buildings data to the package as well. I've just never used them for my work.

Happy to get feedback on that.
