### 2017-11-05 #155 @sckott
Thanks @adamhsparks - Editors are discussing ...
