### 2017-11-06 c661023
Remove DATE from DESCRIPTION
