### 2017-11-06 6154a7c
Clean up vignette output
No need to evaluate the examples. This removes the "Downloaded ..." messages as well as making the installation faster
