### 2017-11-06 c31db42
Enhance package documentation file
