+++ NEWS.md
@@ -1,3 +1,5 @@
+# nasapower 0.1.1.9000
+
 # nasapower 0.1.1
 
 - Fix issues in documentation, typos, incorrect links, etc.
