### 2017-11-06 f8ec772
Update NEWS and static documentation
