### 2017-11-06 6996773
Add new vignette and data to work with
