+++ vignettes/use-case.Rmd
@@ -1,0 +1,122 @@
+---
+title: "Use Case: Using nasapower and the APSIM R Package to Create APSIM .met Files"
+author: "Adam H. Sparks
+"
+date: "`r Sys.Date()`"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{Vignette Title}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+```{r setup, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+```
+
+```{r install_apsim, eval=FALSE}
+install.packages("APSIM")
+```
+
+Load the packages necessary.
+
+```{r load_libraries}
+library("APSIM")
+library("nasapower")
+library("rnaturalearth")
+```
+
+## Get NASA-POWER data
+
+Using the `APSIM::prepareMet()` example where data for Kingsthorpe, Qld,
+is provided by the package, we can compare it with NASA-POWER data for
+Kingsthorpe, Queensland for 2010.
+
+To do that first we need to use `get_cell()` to fetch the data for 2010 and
+create a data frame of weather data which includes the variables, T2M, T2MN,
+T2MX, RAIN and WSM10M.
+
+```{r get_data, eval=TRUE}
+kings_nasa <- get_cell(
+  lonlat = c(151.81, -27.48),
+  stdate = "2010-1-1",
+  endate = "2010-12-31",
+  vars = c("T2M", "T2MN", "T2MX", "RAIN", "WS10M")
+  )
+
+kings_nasa
+```
+
+Now that we have the data...
+
+## Create the .met file
+
+```{r create_met, eval=TRUE}
+# reorder the data columns from nasapower
+kings_met <- kings_nasa[...]
+
+# from the APSIM documentation example
+new_names <-
+  c("Date",
+    "maxt",
+    "mint",
+    "rain",
+    "evaporation",
+    "radn",
+    "vp",
+    "Wind",
+    "RH",
+    "SVP")
+
+units <-
+  c("()",
+    "(oC)",
+    "(oC)",
+    "(mm)",
+    "(mm)",
+    "(MJ/m^2/day)",
+    "()",
+    "()",
+    "()",
+    "()")
+
+prepareMet(kings_met,
+           lat = kings_nasa[1, "LAT"],
+           lon = kings_nasa[1, "LON"],
+           newNames = new_names,
+           units = units)
+```
+
+## Using get_region() with APSIM
+
+To use `get_region()` with `APSIM::prepareMet()` we need to make separate `met`
+files for each cell from the resulting data frame object, which is easy to do
+using _dplyr_.
+
+```{r get_regions, eval=TRUE}
+oz <- get_region(lonlat = c(112.91972, 159.256088, -55.11694, -9.221099),
+                    stdate = "2017-1-1", endate = "2017-1-2")
+```
+
+```{r create_dfs, eval=TRUE}
+library(dplyr)
+
+df <- oz %>%
+ mutate(ids = paste(LON, LAT, sep = '-'))
+
+df2 <- df %>%
+ arrange(ids, LAT)
+
+
+
+ group_by(reg) %>%
+ mutate(n_yrs = length(unique(year))) %>%
+ group_by(reg, ids) %>%
+ summarize(present_yrs = length(unique(year)),
+         all_yrs     = mean(n_yrs, na.rm = T)) %>%
+ filter(present_yrs == all_yrs)
+filter(df, ids %in% df2$ids)
+```
