### 2017-11-06 9db4f07
Update suggests
Add APSIM and dplyr for new use case example
