### 2017-11-06 aa87423
Start to fix issue 1
Incomplete: https://github.com/adamhsparks/nasapower/issues/1
