### 2017-11-06 308c2b7
Start to fix issue 1
Incomplete: https://github.com/adamhsparks/nasapower/issues/1
