### 2017-11-06 e371d43
Fixes Issue 1
