### 2017-11-07 8e339e0
Update use-case.Rmd
Sort and split the region data.frame into a list of data frames for each unique cell in the data
