### 2017-11-07 dfb5d50
Add purrr to suggests
used in use-case vignette
