### 2017-11-07 1602f9a
Edit use-case vignette, don't evaluate R code. Skeleton ready for Keith's input.
