### 2017-11-07 d80cb74
Not using purrr, using lapply in use-case
