+++ DESCRIPTION
@@ -34,7 +34,6 @@ Suggests:
     dplyr,
     ggplot2,
     knitr,
-    purrr,
     rmarkdown,
     rnaturalearth,
     sp,
