### 2017-11-07 17b7458
Further develop region example
