### 2017-11-07 abc273a
Update NEWS.md
