+++ NEWS.md
@@ -1,5 +1,10 @@
 
-# nasapower 0.1.1.9000
+# nasapower 0.1.2
+
+## Bug fixes
+
+- Fixes bug where only first date is reported when using `get_region()` with
+multiple dates. https://github.com/adamhsparks/nasapower/issues/1
 
 ## Minor changes
 
