### 2017-11-07 d09e0f7
Update version number for new release
