+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: nasapower
 Type: Package
 Title: NASA-POWER Agroclimatology Data from R
-Version: 0.1.1.9000
+Version: 0.1.2
 Authors@R: c(person(given = "Adam H.", family = "Sparks",
     role = c("aut", "cre"), email = "adamhsparks@gmail.com",
     comment = "https://orcid.org/0000-0002-0061-8359"),
