### 2017-11-07 9419ef4
Update static documentation
