### 2017-11-07 8cbf0e7
Minor tweaks to ingest data more quickly
