+++ R/get_region.R
@@ -165,7 +165,6 @@ get_region <-
     # Create a data.frame of the NASA - POWER data and add names ---------------
     # Find the immediate prior row to the data, "-END HEADER-"
     min_index <- grep(end, NASA) + 1
-
     max_index <- grep(start, NASA) - 1
     max_index <- max_index[-1]
 
@@ -177,11 +176,14 @@ get_region <-
 
     indices <- data.frame(min_index, max_index)
 
-    indices <- mapply(`:`, indices$min_index, indices$max_index)
+    indices <- data.frame(mapply(`:`, indices$min_index, indices$max_index))
     indices <- unlist(indices, use.names = FALSE)
 
-    NASA <- utils::read.table(textConnection(NASA[indices]),
-                               na.strings = "-")
+    NASA <- utils::read.table(
+      text = NASA[indices],
+      na.strings = "-",
+      nrows = length(indices),
+      stringsAsFactors = FALSE)
 
     # Create a tidy data frame object of lon/lat and data
     NASA <- cbind(location_rows, NASA)
