### 2017-11-07 643af0d
Update NEWS.md
