### 2017-11-07 18d8b1d
Further edits to use case vignette
