+++ DESCRIPTION
@@ -34,6 +34,7 @@ Suggests:
     dplyr,
     ggplot2,
     knitr,
+    lubridate,
     rmarkdown,
     rnaturalearth,
     sp,
