### 2017-11-07 a81670c
Add references to use_case
