+++ vignettes/use-case.Rmd
@@ -14,7 +14,7 @@ vignette: >
 knitr::opts_chunk$set(
   collapse = TRUE,
   comment = "#>",
-  fig.align='center'
+  fig.align = "center"
 )
 ```
 
@@ -33,7 +33,15 @@ library(ggplot2)
 library(dplyr)
 ```
 
-## Get NASA-POWER data
+NASA - POWER data are often used with APSIM (White _et al._ 2012) or other
+models where data may be missing or as a whole data set where data are not
+readily available, _e.g._ modelling rice disease globally (Savary _et al._
+2012).
+
+In this use case we illustrate how to compare NASA-POWER data with station
+data and format for use in APSIM using the R package, APSIM (Fainges 2017).
+
+## Get NASA-POWER Data
 
 Using the `APSIM::prepareMet()` example where data for Kingsthorpe, Qld,
 is provided by the package, we can compare it with NASA-POWER data for
@@ -243,3 +251,20 @@ lapply(1:length(met_list),
        function(i) writeMetFile(met_list[[i]],
                                 fileName = paste0(names(met_list[i]))))
 ```
+
+## References
+
+Justin Fainges (2017) "APSIM: General Utility Functions for the 'Agricultural
+Production Systems Simulator'". R package version 0.9.2.
+<https://CRAN.R-project.org/package=APSIM>
+
+Serge Savary, Andrew Nelson, Laetitia Willocquet, Ireneo Pangga and Jorrel
+Aunario. :Modeling and mapping potential epidemics of rice diseases globally"
+(2012) Crop Protection, Volume 34, Pages 6-17, ISSN 0261-2194 DOI:
+  10.1016/j.cropro.2011.11.009
+  
+
+White, Jeffrey W.; Hoogenboom, Gerrit; Stackhouse, Jr., Paul W.; and Hoell,
+James M., "Evaluation of NASA satellite- and assimilation model-derived
+long-term daily temperature data over the continental US" (2008). NASA
+Publications. 26. <http://digitalcommons.unl.edu/nasapub/26>
