### 2017-11-07 06a7613
Correct citation in references
