### 2017-11-07 af839df
Correct citation in references
