+++ vignettes/use-case.Rmd
@@ -266,7 +266,7 @@ globally" Crop Protection, Volume 34, Pages 6-17, ISSN 0261-2194 DOI:
 10.1016/j.cropro.2011.11.009
 
 Jeffrey W. White, Gerrit Hoogenboom, Paul W. Stackhouse, James M. Hoell (2009)
-Evaluation of NASA satellite- and assimilation model-derived long-term daily
-temperature data over the continental US, In Agricultural and Forest
+"Evaluation of NASA satellite- and assimilation model-derived long-term daily
+temperature data over the continental US" Agricultural and Forest
 Meteorology, Volume 148, Issue 10, 2008, Pages 1574-1584, ISSN 0168-1923, DOI:
 10.1016/j.agrformet.2008.05.017.
