### 2017-11-07 #155 @sckott
### Editor checks:

---

#### Editor comments

Thanks for your submission @adamhsparks

Goodpractice output:

```r
── 0 errors ✔ | 0 warnings ✔ | 0 notes ✔
♥ Aha! Astounding package! Keep up the best work!
```

Seeking reviewers now.

---

Reviewers:  xx xx
Due date: xxx
