### 2017-11-07 cdfb551
minor edits to use-case vignette
