### 2017-11-08 32346b6
Fix bug in get_cell and update documentation
- `get_cell()` had a bug where it would not return all dates, it would return only one.

- Add note to documentation about the order of columns corresponding to order in which vars are listed in function arguments.
