### 2017-11-08 1d0bf6d
Update documentation for get_region()
- Add note to documentation about the order of columns corresponding to order in which vars are listed in function arguments.
