+++ man/get_region.Rd
@@ -60,6 +60,10 @@ meteorological data. The POWER website describes the data as:
   \item{Global coverage on a 1° latitude by 1° longitude grid}
 }
 }
+\note{
+The order in which the `vars` are listed will be the order of the columns in
+the data frame that `get_region()` returns.
+}
 \examples{
 \dontrun{
 # Get POWER data for Australia for 1/1/2017
