### 2017-11-08 e9902a0
Add tests for multiple dates
Test both `get_cell()` and `get_region()` for multiple date handling
