### 2017-11-08 3c3525d
Minor edits to use-case vignette
