### 2017-11-08 5045cb2
Format NEWS.md
