+++ NEWS.md
@@ -1,18 +1,25 @@
+
+--------------------------------------------------------------------------------
+
 # nasapower 0.1.2.9000
 
-# Minor changes
+### Minor improvements
 
 - Optimise arguments used in `read.table` to ingest weather data in the
 `get_cell()` and `get_region()` functions more quickly
 
-# nasapower 0.1.2
+- NEWS now formatted more nicely for easier reading
 
-## Bug fixes
+--------------------------------------------------------------------------------
+
+# nasapower 0.1.2 (2017-11-06)
+
+### Bug fixes
 
 - Fixes bug where only first date is reported when using `get_region()` with
 multiple dates. https://github.com/adamhsparks/nasapower/issues/1
 
-## Minor changes
+### Minor improvements
 
 - Enhanced documentation
 
@@ -26,11 +33,19 @@ multiple dates. https://github.com/adamhsparks/nasapower/issues/1
 
 - Remove DATE from DESCRIPTION
 
-# nasapower 0.1.1
+--------------------------------------------------------------------------------
+
+# nasapower 0.1.1 (2017-11-04)
+
+### Minor improvements
 
 - Fix issues in documentation, typos, incorrect links, etc.
 
-# nasapower 0.1.0
+--------------------------------------------------------------------------------
+
+# nasapower 0.1.0 (2017-11-04)
+
+### New features
 
 * Add new functionality to download regions in addition to single cells
 
@@ -38,15 +53,22 @@ multiple dates. https://github.com/adamhsparks/nasapower/issues/1
 
 * Add startup message
 
+### Minor improvements
+
 * Better documentation
 
-# nasapower 0.0.2
+--------------------------------------------------------------------------------
+
+# nasapower 0.0.2  (2017-11-02)
+
+### New features
 
 * Added citation file
 
-# nasapower 0.0.1
+--------------------------------------------------------------------------------
+
+# nasapower 0.0.1 (2017-11-02)
 
 * Added a `NEWS.md` file to track changes to the package.
 
 * First release, no changes to report yet
-
