### 2017-11-25 5624688
Facet the region maps by YYYYMMDD
