### 2017-11-25 9e316d1
Correct the order of lat/lon values internally for get_region
Fixes https://github.com/adamhsparks/nasapower/issues
