### 2017-11-25 50e39a1
Improve speed with which region data is imported internally and update documentation
