### 2017-11-26 4fa6827
Improve speed with which cell data is imported internally and update documentation
