### 2017-11-26 4d35b6d
Update documentation regarding "var" order
