+++ man/get_region.Rd
@@ -61,8 +61,8 @@ meteorological data. The POWER website describes the data as:
 }
 }
 \note{
-The order in which the `vars` are listed will be the order of the columns in
-the data frame that `get_region()` returns.
+The order in which the \code{vars} are listed will be the order of the
+columns in the data frame that \code{get_region()} returns.
 }
 \examples{
 \dontrun{
