### 2017-11-26 1c3a85b
Clean up R code chunks in vignette
