### 2017-11-26 599b2bc
tidy line lengths in vignette
