+++ vignettes/nasapower.Rmd
@@ -25,7 +25,8 @@ knitr::opts_chunk$set(eval = FALSE,
 _nasapower_ aims to make it quick, easy and efficient to automate downloading
 NASA-POWER agroclimatology data in your R session as a tidy data frame.
 
-_nasapower_ only provides functions, `get_cell()`, for a given 1˚ longitude by 1˚ latitude cell and `get_region()`, both of which will which will download
+_nasapower_ only provides functions, `get_cell()`, for a given 1˚ longitude by
+1˚ latitude cell and `get_region()`, both of which will which will download
 specified variables and return a tidy data frame of the requested data. Weather
 variables can be specified by using the `vars` argument. 
 
@@ -106,7 +107,8 @@ get_cell(lonlat = c(-179.5, -89.5), endate = "1983-01-02", vars = "WS10M")
 ## Using get_region()
 
 The `get_region()` function accepts the same arguments as `get_cell()`, however,
-the `lonlat` value is a length-4 numeric vector of the minimum longitude, maximum longitude, minimum latitude and maximum latitude.
+the `lonlat` value is a length-4 numeric vector of the minimum longitude,
+maximum longitude, minimum latitude and maximum latitude.
 
 This function is useful for retrieving whole states, countries or global
 surfaces. This example illustrates fetching and graphing the average temperature
