### 2017-11-26 62d2d53
Fix documentation
- `lat` was still referred to in README and vignette as an argument. It was removed for `lonlat`

- Add code to show vignette in R session
