+++ README.Rmd
@@ -33,7 +33,7 @@ using the following code:
 
 ```{r, eval=FALSE}
 # install.packages("devtools", dep = TRUE)
-devtools::install_github("adamhsparks/nasapower")
+devtools::install_github("adamhsparks/nasapower", build_vingettes = TRUE)
 ```
 
 ## Introduction
@@ -47,7 +47,7 @@ variables can be specified by using the `vars` argument.
 
 Both functions accept four arguments.
 
-```r
+```{r arguments, eval=FALSE}
 get_cell(lonlat = NULL, vars = c("T2M", "T2MN", "T2MX", "RH2M"),
   stdate = "1983-1-1", endate = Sys.Date())
 ```
@@ -55,8 +55,6 @@ get_cell(lonlat = NULL, vars = c("T2M", "T2MN", "T2MX", "RH2M"),
 - `lonlat` for `get_cell()` A length-2 numeric vector giving the decimal degree longitude and and latitude in that order for cell data to download
 
 - `lonlat` for `get_region()` A length-4 numeric vector of the minimum longitude, maximum longitude, minimum latitude and maximum latitude.
-
-- `lat` Latitude value to query (single value only)
 
 - `vars` Weather variables to query for download. Valid `vars` include:
 
@@ -96,7 +94,17 @@ library(nasapower)
 get_cell(lonlat = c(-179.5, -89.5))
 ```
 
-More documentation is available in the vignette, <https://adamhsparks.github.io/nasapower/articles/nasapower.html>.
+## Documentation
+
+More documentation is available in the vignette in your R session.
+
+```{r vignette, eval=FALSE}
+vignette("nasapower")
+``` 
+
+Or available online.
+
+<https://adamhsparks.github.io/nasapower/articles/nasapower.html>.
 
 ## Use of POWER Data
 
