### 2017-11-26 3fdcfef
Correct spelling of "vignette"
