### 2017-11-26 cd114be
Edit documentation section to read more easily
