### 2017-11-27 d0b3390
Update NEWS.md
