### 2017-11-27 #155 @UniversalTourist
>##### For packages co-submitting to JOSS
>
>- [ ] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

Estimated hours spent reviewing: It took almost, totally, 4-5 hours to check everything. During that time I studied what is NASA-POWER, agroclimatology and read rOpenSci guidelines, learned about Travis-CI and running tests.

---

This is the first time I heard about NASA-POWER and it is a great data source for Earth scientists. These package provides a data frame output and it will make the data preparation process more easier than copy-pasting.
(I have tested functions for region of my country and already sent the package's Github link for my colleges :) )

- Installation:
With MacOS:
I used macOS Sierra, Verison 10.12 and I was able to install the package from Github without any problem by devtools.

- Code Overall:
These package has two functions to get data for a cell or for a region. Both functions has nice structered, easy to follow and also nice explained with comments, I enjoyed reviewing this part especially. (I also learned about how httr works by running these codes step by step.)

- rOpenSci criteria:
Package name is really nice. Function and variable naming is proper. Licence and Code of Conduct are available.

- Readme:
Readme.rmd is clear and easy to follow steps for installations.

- Vignette:
There are more details and examples about other function. Also the examples provide more information and help to illustrate the importance of the package.
In the introduction part there is only one mistake, the "which will" words are written for two times: "both of which will which will download".
I notified author about plots on vignette which seems different when I run on my system, and problem was resolved with [a commit](https://github.com/adamhsparks/nasapower/commit/62d2d538deb711cd71107374e35d134cf67311bf) fastly when I contacted with the author.

- Errors and console messages:
They are all clear.

- R CMD check:
R CMD check passes without any warnings or any errors.
- [X] 0 errors | 0 warnings | 0 notes
:)

- Other comments
NASA-POWER provides data at a resolution of 1˚ longitude by 1˚ latitude, which is good enough most of the time. Maybe you can also write about other data sources which provide higher resolutions in the package readme file for the people who come to this package repo for higher resolution.

p.s This is my first time at reviewing a package and I hope I can be helpful :)
