### 2017-11-27 d1492c1
Update NEWS.md
