### 2017-11-27 63986a3
Update devtools package installation method and update static documentation site
