### 2017-11-27 #155 @adamhsparks
Thank you @UniversalTourist

I have fixed the vignette text you noted where it said "which will which will", https://github.com/adamhsparks/nasapower/commit/6dd0d9cfd7d3dd71bc0cd435be52cda6c7a2a91e.
