### 2017-11-27 8b8a11a
Update README
- Fix list md formatting

- Add link to the NOAA list of gridded data sets
