+++ README.Rmd
@@ -45,7 +45,8 @@ devtools::install_github("adamhsparks/nasapower", build_vignettes = TRUE)
 _nasapower_ aims to make it quick, easy and efficient to automate downloading
 NASA-POWER agroclimatology data in your R session as a tidy data frame.
 
-_nasapower_ only provides functions, `get_cell()`, for a given 1˚ longitude by 1˚ latitude cell and `get_region()`, both of which will which will download
+_nasapower_ only provides functions, `get_cell()`, for a given 1˚ longitude by
+1˚ latitude cell and `get_region()`, both of which will which will download
 specified variables and return a tidy data frame of the requested data. Weather
 variables can be specified by using the `vars` argument. 
 
@@ -56,9 +57,11 @@ get_cell(lonlat = NULL, vars = c("T2M", "T2MN", "T2MX", "RH2M"),
   stdate = "1983-1-1", endate = Sys.Date())
 ```
 
-- `lonlat` for `get_cell()` A length-2 numeric vector giving the decimal degree longitude and and latitude in that order for cell data to download
+- `lonlat` for `get_cell()` A length-2 numeric vector giving the decimal degree
+longitude and and latitude in that order for cell data to download
 
-- `lonlat` for `get_region()` A length-4 numeric vector of the minimum longitude, maximum longitude, minimum latitude and maximum latitude.
+- `lonlat` for `get_region()` A length-4 numeric vector of the minimum
+longitude, maximum longitude, minimum latitude and maximum latitude.
 
 - `vars` Weather variables to query for download. Valid `vars` include:
 
@@ -68,15 +71,19 @@ get_cell(lonlat = NULL, vars = c("T2M", "T2MN", "T2MX", "RH2M"),
 
     * lwv_dwn - Average downward longwave radiative flux (MJ/m^2/day)
 
-    * T2M - Average air temperature at 2m above the surface of the Earth (degrees C)
+    * T2M - Average air temperature at 2m above the surface of the Earth
+    (degrees C)
 
-    * T2MN - Minimum air temperature at 2m above the surface of the Earth (degrees C)
+    * T2MN - Minimum air temperature at 2m above the surface of the Earth
+    (degrees C)
 
-    * T2MX - Maximum air temperature at 2m above the surface of the Earth (degrees C)
+    * T2MX - Maximum air temperature at 2m above the surface of the Earth
+    (degrees C)
 
     * RH2M - Relative humidity at 2m above the surface of the Earth (%)
 
-    * DFP2M - Dew/Frost point temperature at 2m above the surface of the Earth (degrees C)
+    * DFP2M - Dew/Frost point temperature at 2m above the surface of the Earth
+    (degrees C)
 
     * RAIN - Average precipitation (mm/day)
 
@@ -117,10 +124,20 @@ Applied Science Program."
 ## Meta
 
 * Please [report any issues or bugs](https://github.com/adamhsparks/nasapower/issues).
+
 * License: MIT
-* Get citation information for `nasapower` in R doing `citation(package = "nasapower")`
-* Please note that this project is released with a [Contributor Code of Conduct](CONDUCT.md).
+
+* Get citation information for `nasapower` in R by typing
+`citation(package = "nasapower")`
+
+* Please note that this project is released with a
+[Contributor Code of Conduct](CONDUCT.md).
 By participating in this project you agree to abide by its terms.
+
+* The U.S. Earth System Research Laboratory, Physical Science Division of the
+National Atmospheric & Oceanic Administration (NOAA) maintains a list of
+Gridded Climate Data sets that provide different data and different resolutions
+<https://www.esrl.noaa.gov/psd/data/gridded/>
 
 ## References
 <https://power.larc.nasa.gov>
