### 2017-11-28 c09784a
Default to download all available weather data in both functions
Change behaviour and update documentation to match
