+++ R/get_cell.R
@@ -18,8 +18,8 @@
 #' }
 #' @param lonlat A length-2 numeric vector giving the decimal degree longitude
 #' and latitude in that order for cell data to download
-#' @param vars Weather variables to download, defaults to T2M, T2MN, T2MX and
-#' RH2m. Valid variables are:
+#' @param vars Weather variables to download, defaults to all available.
+#' Valid variables are:
 #' \itemize{
 #' \item{\strong{toa_dwn} - Average top-of-atmosphere insolation (MJ/m^2/day)}
 #' \item{\strong{swv_dwn} - Average insolation incident on a horizontal surface
@@ -66,7 +66,13 @@ get_cell <-
            vars = c("T2M",
                     "T2MN",
                     "T2MX",
-                    "RH2M"),
+                    "RH2M",
+                    "toa_dwn",
+                    "swv_dwn",
+                    "lwv_dwn",
+                    "DFP2M",
+                    "RAIN",
+                    "WS10M"),
            stdate = "1983-1-1",
            endate = Sys.Date()) {
     if (is.null(lonlat) | length(lonlat) != 2 | !is.numeric(lonlat)) {
