### 2017-11-28 3cbf7db
Update vignette and static documentation
Reflect change in behaviour that ALL variables are downloaded by default
