### 2017-11-28 af47360
Correct documentation example
