+++ README.Rmd
@@ -97,8 +97,8 @@ earlier data)
 
 ### Basic example of using get_cell()
 
-Fetch T2M, T2MN, T2MX and RH2M for 1983-1-1 to current date for the cell at
-longitude -179.5 and latitude -89.5.
+Fetch all weather variables available for 1983-1-1 to current date for the cell
+at longitude -179.5 and latitude -89.5.
 
 ```{r basic_example, eval=FALSE}
 library(nasapower)
