### 2017-12-11 fb05709
Add hex sticker
