+++ README.Rmd
@@ -15,7 +15,8 @@ library(nasapower)
 
 ## Introduction
 
-_nasapower_ aims to make it quick and easy to automate downloading
+<img align="right" src="man/figures/nasapower-hex.png"> _nasapower_ aims to make
+it quick and easy to automate downloading 
 [NASA-POWER](https://power.larc.nasa.gov) agroclimatology data in your R session
 as a tidy data frame for agricultural analysis and use in modelling or other
 purposes. POWER (Prediction Of Worldwide Energy Resource) data are freely
