### 2017-12-12 a3bc924
Fix bug reports URL in DESCRIPTION
