### 2017-12-12 f213c81
Clean up documentation
