+++ README.Rmd
@@ -17,7 +17,6 @@ library(nasapower)
 
 <img align="right" src="man/figures/nasapower-hex.png"> _nasapower_ aims to make
 it quick and easy to automate downloading
->>>>>>> master
 [NASA-POWER](https://power.larc.nasa.gov) agroclimatology data in your R session
 as a tidy data frame for agricultural analysis and use in modelling or other
 purposes. POWER (Prediction Of Worldwide Energy Resource) data are freely
