### 2017-12-12 f30ad7a
Update documentation and formatting
