+++ R/nasapower-package.R
@@ -1,5 +1,5 @@
 
-#' Downloads NASA-POWER Agroclimatology Variables and Returns a Tidy Data Frame
+#' NASA-POWER Agroclimatology Data from R
 #'
 #' @name nasapower
 #' @author Adam H Sparks
@@ -17,12 +17,14 @@
 #' While \code{nasapower} does not redistribute the data in anyway, we
 #' encourage users to follow the requests of the POWER Project Team.
 #'
+#'\preformatted{
 #' When POWER data products are used in a publication, we request the
 #' following acknowledgment be included:
 #'
 #' "These data were obtained from the NASA Langley Research Center
 #' POWER Project funded through the NASA Earth Science Directorate Applied
 #' Science Program."
+#' }
 #'
 #' @seealso
 #' \itemize{
