### 2017-12-12 #155 @alisonboyer
Package Review

As the reviewer I confirm that there are no conflicts of interest for me to review this work (If you are unsure whether you are in conflict, please speak to your editor before starting your review).

- [x] A statement of need clearly stating problems the software is designed to solve and its target audience in README
- [x] Installation instructions: for the development version of package and any non-standard dependencies in README
- [x] Vignette(s) demonstrating major functionality that runs successfully locally
- [x] Function Documentation: for all exported functions in R help
- [x] Examples for all exported functions in R Help that run successfully locally
- [x] Community guidelines including contribution guidelines in the README or CONTRIBUTING, and DESCRIPTION with URL, BugReports and Maintainer (which may be autogenerated via Authors@R).

- [x]  Installation: Installation succeeds as documented.
- [x]  Functionality: Any functional claims of the software been confirmed.
- [ ]  Performance: Any performance claims of the software been confirmed.
Comment: How large of a region can you request using this service? Some notes about the expected size of the data frame returned would be helpful. For example, a 2 x 4 degree region with 3 variables and all available time steps was ~ 2MB. If I request the entire globe, what will happen?
- [x]  Automated tests: Unit tests cover essential functions of the package and a reasonable range of inputs and conditions. All tests pass on the local machine.
Comment: I do not know how to run unit tests. My local tests of the basic functionality all ran flawlessly.
- [x]  Packaging guidelines: The package conforms to the rOpenSci packaging guidelines

Review Comments:

Providing a recommended citation for the data is very nice. Does the data have a DOI?
The link to other climate data provided by NOAA ESRL is helpful.
