### 2017-12-12 ebe5111
modify travis.yml to correct error
