### 2017-12-13 3bab17e
Update documentation
Add statement about performance. How long it might take to download a global surface for one day and the size in GB for 365 global surfaces in a data frame.

Also add links to split, list and rasterFromXYZ documentation
