### 2017-12-13 #155 @adamhsparks
Thank you, @alisonboyer.

I didn't have a statement about performance because it will vary by the data requested and Internet connection. However, I ran an example this morning and added a statement to the vignette in this commit, https://github.com/adamhsparks/nasapower/commit/3bab17ecc38f42d8998042c5866503f5163279e3 about what some possible realistic times and data frame sizes could be for this.

AFAIK there is no DOI for the POWER data. As you noted, I've included their statement about how they wish to be acknowledged and the link to https://power.larc.nasa.gov/documents/Agroclimatology_Methodology.pdf is included in the documentation in several locations as well.
