### 2017-12-13 #155 @adamhsparks
For both @UniversalTourist  and @alisonboyer, I've also expanded the documentation in the vignette to provide an example of how to turn the data frame in to a `raster::brick` spatial object for users that need it to be a spatial object not a normal data frame.
