### 2017-12-13 58d71c7
Add raster to suggests
raster is now used in an example in the vignette
