### 2017-12-13 b0566ac
Fix capitalisation in text
