+++ README.Rmd
@@ -138,7 +138,7 @@ By participating in this project you agree to abide by its terms.
 
 * The U.S. Earth System Research Laboratory, Physical Science Division of the
 National Atmospheric & Oceanic Administration (NOAA) maintains a list of
-Gridded Climate Data sets that provide different data and different resolutions
+gridded climate data sets that provide different data and different resolutions
 <https://www.esrl.noaa.gov/psd/data/gridded/>
 
 ## References
