### 2017-12-13 e1a6b3b
Update NEWS
