### 2017-12-13 a4ab7c9
Update NEWS
