### 2017-12-13 76947ad
Update NEWS
