### 2017-12-13 4b85438
Add X-schema.org info to DESCRIPTION
