+++ DESCRIPTION
@@ -39,3 +39,5 @@ Suggests:
     sp,
     viridis
 VignetteBuilder: knitr
+X-schema.org-applicationCategory: Tools
+X-schema.org-keywords: NASA, meteorological-data, weather, global, weather, weather-data, meteorology, NASA-POWER, agroclimatology, earth-science, data-access
