### 2017-12-13 94e2440
Add functionality to check if server is responding
If server is not responding, stop and issue error message to user.
