### 2017-12-13 #155 @adamhsparks
This afternoon the POWER webpage stopped serving data. I've added a check now that pings the server first, five times to see if it's responding, before proceeding with the request.

If the reviewers could check this as it was added after both reviews?

https://github.com/adamhsparks/nasapower/commit/94e24406fff00f52db1d2c3346c78c115e7ab7b7
