### 2017-12-13 dbcd72f
clean up url in durl
Use `url` object used for checking server responses to clean up the code.

No need to repeat
