### 2017-12-13 #155 @adamhsparks
@UniversalTourist, please wait. Now that it's responding again, it's clear that my fix didn't work. I need to find a new way.
