### 2017-12-13 d464fa9
Revert "Add functionality to check if server is responding"
This reverts commit 94e24406fff00f52db1d2c3346c78c115e7ab7b7.
