+++ R/get_region.R
@@ -110,9 +110,7 @@ get_region <-
 
     durl <-
       paste0(
-        "https://",
-        url,
-        "&ms=",
+        "https://power.larc.nasa.gov/cgi-bin/agro.cgi?&ms=",
         format(as.Date(stdate), "%m"),
         "&area=area&",
         "latmin=",
@@ -133,7 +131,7 @@ get_region <-
         format(as.Date(endate), "%m"),
         "&de=",
         format(as.Date(endate), "%d"),
-        "&submit=Submit&lonmin=",
+        "&email=agroclim40larc.nasa.gov&submit=Submit&lonmin=",
         lonlat[1]
       )
 
