### 2017-12-13 2c30cec
Revert "Revert "Add functionality to check if server is responding""
This reverts commit d464fa93b751cc4c68e288910ab0021c0b96257b.
