### 2017-12-13 6faecaa
Fix internal_functions to check POWER response
