+++ R/internal_functions.R
@@ -3,7 +3,7 @@
 #' @noRd
 .check_response <- function(url) {
 
-  if (all(is.na(pingr::ping(url, count = 5)))) {
+  if (httr::http_status(httr::GET(url))$category != "Success") {
     stop("\nThe POWER website does not appear to be responding.\n",
          "Please try again later.\n")
   }
