### 2017-12-13 1d8b28a
remove pingr
