### 2017-12-13 d5407ff
Fix website response check and tests
