### 2017-12-13 #155 @adamhsparks
OK, fixed now with a new better way of doing checking whether the NASA - POWER site is serving data or not.
