### 2017-12-13 cd1401c
Shorten line lengths
