+++ R/get_region.R
@@ -178,7 +178,9 @@ get_region <-
 
     # add duplicate rows for n dates
     location_rows <- location_rows[rep(row.names(location_rows),
-                                       each = as.numeric((endate - stdate) + 1)),
+                                       each = as.numeric(
+                                         (endate - stdate) + 1)
+                                       ),
                                    1:2]
 
     location_rows <-
