### 2017-12-13 66ebf32
Fix comment in .check_response test
