+++ DESCRIPTION
@@ -4,7 +4,7 @@ Title: NASA-POWER Agroclimatology Data from R
 Version: 0.1.2.9000
 Authors@R: c(person(given = "Adam H.", family = "Sparks",
     role = c("aut", "cre"), email = "adamhsparks@gmail.com",
-    comment = "0000-0002-0061-8359"),
+    comment = c(ORCID = "0000-0002-0061-8359")),
     person(given = "Jorrel Khalil S.", family = "Aunario",
     role = "ctb", email = "jaunario@gmail.com"))
 URL: https://github.com/adamhsparks/nasapower
