### 2017-12-17 #155 @UniversalTourist
Everything seems OK! Spatial objects example will really make this package more attractive! 💯  And I love the sticker! 😄
