### 2017-12-18 6b72e42
Fix description for `get_region()`
incorrectly stated that it downloaded data for a 1 x 1 degree cell
