+++ R/get_region.R
@@ -2,8 +2,8 @@
 #' Download NASA-POWER Agroclimatology Variables for a Given Region and Return a Tidy Data Frame
 #'
 #' @description Download NASA-POWER (Prediction of Worldwide Energy Resource)
-#' agroclimatology data for a region given 1 degree x 1 degree cell for a range
-#' of years.
+#' agroclimatology data for a region given the four corners of the region of
+#' interest for a range of years.
 #'
 #' The agroclimatology data are satellite and modelled derived solar and
 #' meteorological data. The POWER website describes the data as:
