+++ man/get_region.Rd
@@ -46,8 +46,8 @@ for the requested region.
 }
 \description{
 Download NASA-POWER (Prediction of Worldwide Energy Resource)
-agroclimatology data for a region given 1 degree x 1 degree cell for a range
-of years.
+agroclimatology data for a region given the four corners of the region of
+interest for a range of years.
 
 The agroclimatology data are satellite and modelled derived solar and
 meteorological data. The POWER website describes the data as:
