### 2017-12-18 df48fd8
Add new checks for lon/lat values and tests for checks
