+++ R/get_region.R
@@ -81,13 +81,11 @@ get_region <-
            stdate = "1983-1-1",
            endate = Sys.Date()) {
 
+    lonlat <- .check_lonlat_region(lonlat)
+
     url <-
       "power.larc.nasa.gov/cgi-bin/agro.cgi?email=agroclim@larc.nasa.gov"
     .check_response(url)
-
-    if (is.null(lonlat) | length(lonlat) != 4 | !is.numeric(lonlat)) {
-      stop("lonlat must be provided in a length-4 numeric vector.\n")
-    }
 
     stdate <- as.Date(stdate)
     endate <- as.Date(endate)
