### 2017-12-18 29513e6
Update NEWS.md
