### 2017-12-18 3b49370
Add JOSS paper draft
