### 2017-12-18 6297387
Fix bugs with retrieving POWER data
