+++ R/get_cell.R
@@ -76,7 +76,7 @@ get_cell <-
            stdate = "1983-1-1",
            endate = Sys.Date()) {
 
-    lonlat <- .check_lonlat_cell(lonlat)
+    .check_lonlat_cell(lonlat)
 
     url <-
       "power.larc.nasa.gov/cgi-bin/agro.cgi?email=agroclim@larc.nasa.gov"
@@ -112,14 +112,14 @@ get_cell <-
         url,
         "&p=",
         download_vars,
-        "lat=",
+        "lon=",
+        lonlat[1],
+        "&lat=",
         lonlat[2],
         "&ye=",
         format(as.Date(endate), "%Y"),
         "&me=",
         format(as.Date(endate), "%m"),
-        "&lon=",
-        lonlat[1],
         "&submit=Submit&ms=",
         format(as.Date(stdate), "%m"),
         "&step=1&de=",
@@ -133,12 +133,6 @@ get_cell <-
     # Read lines from the NASA-POWER website
     NASA <-
       httr::content(httr::GET(durl, httr::progress()), encoding = "UTF8")
-
-    # Check contents to be sure that it is data and not an error message
-
-    if (grepl("No location", NASA)) {
-      stop("You have provided invalid ")
-    }
 
     # clear console
     message("\n")
