### 2017-12-18 a4bc4a4
Remove use-case
