### 2017-12-18 e2e3f3d
Clean up URLs
Put URL code into a reasonable order that's more easily human readable and debugable
