### 2017-12-18 4f4c020
Bit more work on the paper
