### 2017-12-18 3fba538
Cleanup download vars
Remove the last "&p=" rather than only removing "&p" only.
