### 2017-12-18 b740651
Punctuate user messages
