### 2017-12-18 b1d4b36
Update NEWS.md
