### 2017-12-18 64840ce
Update NEWS
