### 2017-12-18 fd2c151
Clean up JOSS paper
