### 2017-12-18 f3970ff
Clean up extra line at line 1
