+++ vignettes/use-case.Rmd
@@ -1,0 +1,265 @@
+---
+title: "Use Case: Using nasapower and the APSIM R Package to Create APSIM .met Files"
+author: "Adam H. Sparks
+"
+date: "`r Sys.Date()`"
+output:
+  rmarkdown::html_vignette:
+        toc: true
+vignette: >
+  %\VignetteIndexEntry{Vignette Title}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+```{r setup, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>",
+  fig.align = "center"
+)
+```
+
+```{r install_apsim, eval=FALSE}
+install.packages("APSIM")
+```
+
+Load the packages necessary.
+
+```{r load_libraries, message=FALSE}
+library(APSIM)
+library(nasapower)
+library(rnaturalearth)
+library(lubridate)
+library(ggplot2)
+library(dplyr)
+```
+
+NASA - POWER data are often used with APSIM (White _et al._ 2012) or other
+models where data may be missing or as a whole data set where data are not
+readily available, _e.g._ modelling rice disease globally (Savary _et al._
+2012).
+
+In this use case we illustrate how use NASA-POWER radiation data with existing
+station data and format for use in APSIM using the R package, APSIM
+(Fainges 2017) and how it can be used for a larger modelling exercise like
+the EPIRICE paper (Savary _et al._ 2012)
+
+## Get NASA-POWER Data
+
+Using the `APSIM::prepareMet()` example where data for Kingsthorpe, Qld,
+is provided by the package, we can compare it with NASA-POWER radiation for
+Kingsthorpe, Queensland for 2010.
+
+To do that first we need to use `get_cell()` to fetch the data necessary to
+calculate SRAD for 2010.
+
+```{r get_data_show, eval=FALSE}
+kings_nasa <- get_cell(
+  lonlat = c(151.81, -27.48),
+  stdate = "2010-1-1",
+  endate = "2010-12-31",
+  vars = c("T2MX", "T2MN", "RAIN")
+  )
+```
+
+```{r get_data_execute, eval=TRUE, message=FALSE, include=FALSE}
+# This code chunk actually downloads the data, but it will show
+# download progress, so we don't include it.
+kings_nasa <- get_cell(
+  lonlat = c(151.81, -27.48),
+  stdate = "2010-1-1",
+  endate = "2010-12-31",
+  vars = c("T2MX", "T2MN", "RAIN")
+  )
+```
+
+## Calculate SRAD
+
+```{r}
+
+```
+
+## Compare the data
+
+To compare the data more easily using graphs we'll join the two data frames
+using _dplyr_. 
+
+First convert the Date column from kingsData to a Date object using _lubridate_.
+
+```{r conv_date, eval=TRUE}
+kingsData$Date <- lubridate::dmy(kingsData$Date)
+```
+
+Next, join the two data frames together to create a data frame called,
+`compare`.
+
+```{r compare, eval=TRUE}
+compare <- left_join(kingsData, kings_nasa, by = c("Date" = "YYYYMMDD"))
+```
+
+### Graphing the data
+
+Graph maximum temperature, comparing `kingsData` to NASA-POWER data and check
+correlation between the values.
+
+### Maximum Temperature
+
+```{r graph_maxt, eval=TRUE}
+ggplot(compare, aes(x = max.temp, y = T2MX)) +
+  geom_point(colour = "red") +
+  theme_minimal()
+
+cor(compare$max.temp, compare$T2MX)
+```
+
[…]
