+++ vignettes/use-case.Rmd
@@ -29,10 +29,7 @@ Load the packages necessary.
 ```{r load_libraries, message=FALSE}
 library(APSIM)
 library(nasapower)
-library(rnaturalearth)
-library(lubridate)
 library(ggplot2)
-library(dplyr)
 ```
 
 NASA - POWER data are often used with APSIM (White _et al._ 2012) or other
@@ -41,117 +38,85 @@ readily available, _e.g._ modelling rice disease globally (Savary _et al._
 2012).
 
 In this use case we illustrate how use NASA-POWER radiation data with existing
-station data and format for use in APSIM using the R package, APSIM
-(Fainges 2017) and how it can be used for a larger modelling exercise like
-the EPIRICE paper (Savary _et al._ 2012)
+station data that lacks radiation and format for use in APSIM by creating a
+.met file using the R package, _APSIM_ (Fainges 2017).
 
-## Get NASA-POWER Data
+## Download NASA-POWER Solar Radiation Data
 
 Using the `APSIM::prepareMet()` example, where data for Kingsthorpe, Qld,
 is provided by the package, we can create an equivalent data set from the 
 NASA-POWER data compare it with the supplied data for Kingsthorpe, Queensland
-for 2010. As a proxy for potential incoming radiation, SRAD, we'll use
-`toa_dwn`, top-of-atmosphere insolation, from NASA - POWER.
+location 151.81 degrees longitude and -27.48 degrees latitude, for 2010. As a
+proxy for potential incoming radiation, `SRAD`, we'll use `swv_dwn`,
+top-of-atmosphere insolation, from NASA - POWER.
 
 ```{r get_data_show, eval=FALSE}
-kings_nasa <- get_cell(
+swv_dwn <- get_cell(
   lonlat = c(151.81, -27.48),
   stdate = "2010-1-1",
   endate = "2010-12-31",
-  vars = c("T2MX", "T2MN", "RAIN", "toa_dwn")
+  vars = "swv_dwn"
   )
 ```
 
 ```{r get_data_execute, eval=TRUE, message=FALSE, include=FALSE}
 # This code chunk actually downloads the data, but it will show
 # download progress, so we don't include it.
-kings_nasa <- get_cell(
+swv_dwn <- get_cell(
   lonlat = c(151.81, -27.48),
   stdate = "2010-1-1",
   endate = "2010-12-31",
-  vars = c("T2MX", "T2MN", "RAIN", "toa_dwn")
+  vars = "swv_dwn"
   )
 ```
 
-## Calculate incoming solar radiation from toa_dwn
 
-```{r incoming_rad, , eval=TRUE, message=FALSE, include=TRUE}
+## Compare the Data
 
-```
-
-## Compare the data
-
-To compare the data more easily using graphs we'll join the two data frames
-using _dplyr_. 
-
-First convert the Date column from kingsData to a Date object using _lubridate_.
-
-```{r conv_date, eval=TRUE}
-kingsData$Date <- lubridate::dmy(kingsData$Date)
-```
-
-Next, join the two data frames together to create a data frame called,
-`compare`.
+To compare the data more easily using a graph and correlation measures, join the
+two data frames to create a data frame called, `compare`.
 
 ```{r compare, eval=TRUE}
-compare <- left_join(kingsData, kings_nasa, by = c("Date" = "YYYYMMDD"))
+compare <- data.frame(kingsData$rad, swv_dwn$swv_dwn)
 ```
 
-### Graphing the data
+### Graph the Data
 
-Graph maximum temperature, comparing `kingsData` to NASA-POWER data and check
-correlation between the values.
-
-### Maximum Temperature
-
[…]
