### 2017-12-19 00d92a4
Update paper to include EPIRICE reference
