+++ inst/paper/paper.bib
@@ -1,91 +1,113 @@
 @article{bai2010evaluation,
-  Author = {Bai, Jinshun and Chen, Xinping and Dobermann, Achim and Yang, Haishun and Cassman, Kenneth G and Zhang, Fusuo},
-  Journal = {Agronomy Journal},
-  Number = {1},
-  Pages = {9--16},
-  Publisher = {American Society of Agronomy},
-  Title = {Evaluation of NASA satellite-and model-derived weather data for simulation of maize yield potential in China},
-  Volume = {102},
-  Year = {2010}}
+	Author = {Bai, Jinshun and Chen, Xinping and Dobermann, Achim and Yang, Haishun and Cassman, Kenneth G and Zhang, Fusuo},
+	Journal = {Agronomy Journal},
+	Number = {1},
+	Pages = {9--16},
+	Publisher = {American Society of Agronomy},
+	Title = {Evaluation of NASA satellite-and model-derived weather data for simulation of maize yield potential in China},
+	Volume = {102},
+	Year = {2010}}
 
 @manual{hijmans2017,
-  Author = {Robert J. Hijmans},
-  Note = {R package version 2.6-7},
-  Title = {{raster}: Geographic Data Analysis and Modeling},
-  Url = {https://CRAN.R-project.org/package=raster},
-  Year = {2017}}
-
-@manual{sparks2017,
-  Author = {Adam H. Sparks},
-  Note = {R package version 0.1.3},
-  Title = {{nasapower}: NASA-POWER Agroclimatology Data from R},
-  Url = {https://CRAN.R-project.org/package=nasapower},
-  Year = {2017}}
-
-@article{vanwart2013,
-  Author = {{van Wart}, Justin and Grassini, Patricio and Cassman, Kenneth G.},
-  Doi = {10.1111/gcb.12302},
-  Issn = {1365-2486},
-  Journal = {Global Change Biology},
-  Keywords = {crop model, maize, rice, weather data, wheat, yield potential},
-  Number = {12},
-  Pages = {3822--3834},
-  Title = {Impact of derived global weather data on simulated crop yields},
-  Url = {http://dx.doi.org/10.1111/gcb.12302},
-  Volume = {19},
-  Year = {2013}}
-
-@article{vanwart2015,
-  Author = {Justin {van Wart} and Patricio Grassini and Haishun Yang and Lieven Claessens and Andrew Jarvis and Kenneth G. Cassman},
-  Doi = {https://doi.org/10.1016/j.agrformet.2015.02.020},
-  Issn = {0168-1923},
-  Journal = {Agricultural and Forest Meteorology},
-  Keywords = {Yield potential, Yield variability, Weather data, Crop simulation model},
-  Number = {Supplement C},
-  Pages = {49 - 58},
-  Title = {Creating long-term weather data from thin air for crop simulation modeling},
-  Url = {http://www.sciencedirect.com/science/article/pii/S0168192315000696},
-  Volume = {209-210},
-  Year = {2015}}
+	Author = {Robert J. Hijmans},
+	Note = {R package version 2.6-7},
+	Title = {{raster}: Geographic Data Analysis and Modeling},
+	Url = {https://CRAN.R-project.org/package=raster},
+	Year = {2017},
+	Bdsk-Url-1 = {https://CRAN.R-project.org/package=raster}}
 
 @article{ojeda2017,
-  Author = {Ojeda, Jonathan J and Volenec, Jeffrey J and Brouder, Sylvie M and Caviglia, Octavio P and Agnusdei, M{\'o}nica G},
-  Journal = {GCB Bioenergy},
[…]
