### 2017-12-19 cb29e0a
Add APSIM Keating citation
