+++ vignettes/use-case.Rmd
@@ -174,11 +174,11 @@ Production Systems Simulator'". R package version 0.9.2.
 
 Keating, Brian A., Peter S. Carberry, Graeme L. Hammer, Mervyn E. Probert,
 Michael J. Robertson, D. Holzworth, Neil I. Huth et al. "An overview of APSIM,
-a model designed for farming systems simulation." European journal of agronomy
+a model designed for farming systems simulation." European Journal of Agronomy
 18, no. 3 (2003): 267-288.
 
 Serge Savary, Andrew Nelson, Laetitia Willocquet, Ireneo Pangga and Jorrel
-Aunario (2012) "Modeling and mapping potential epidemics of rice diseases
+Aunario (2012) "Modeling and Mapping potential epidemics of rice diseases
 globally" Crop Protection, Volume 34, Pages 6-17, ISSN 0261-2194 DOI:
 10.1016/j.cropro.2011.11.009
 
