### 2017-12-19 f3089ff
Update NEWS.md
