### 2017-12-19 827da45
Few more minor formatting fixes in NEWS
