### 2017-12-19 #155 @adamhsparks
@UniversalTourist and @alisonboyer, I've added a use-case vignette now, https://adamhsparks.github.io/nasapower/articles/use-case.html. And I've included further validation of user-entered lon/lat values for both functions, https://github.com/adamhsparks/nasapower/commit/df48fd8128a242e084145052eb63c1e538595ff4. I hope that I've covered off on all the requested changes (and then some).

@sckott, I've started drafting a JOSS paper. I was waiting to see if there were any major changes requested by the reviewers before writing it.
