### 2017-12-19 865b685
Remove "pretty good" and use "strong"
