### 2017-12-19 88f5fa5
Still wordsmithing the use-case
