### 2017-12-19 fd9b928
Edits to paper.md
