### 2017-12-19 356024c
Fix YAML in paper.md
