### 2017-12-19 a48175a
fix numbering in inline list
