### 2017-12-19 #155 @sckott
@adamhsparks If okay with you, i want to take a quick look at the latest version to see if there's anything to comment on ...
