### 2017-12-20 #155 @sckott
some thoughts:

for both `get_cell` and `get_region`

- unless you have some edge case stuff, would it not be easier to construct a named list of query params to pass to `GET`?
- seems smart to break apart the `GET` and `content` calls on separate lines, w/ a `stop_for_status()` or similar in between them so users get nice error messages when a server or client error happens
- I think there's no check on the `vars` param internally, so e.g. ,

```r
get_cell(lonlat = c(-179.5, -89.5), vars = 'asdfasdf')
#> Error in `[.data.frame`(NASA, , c(refcols, setdiff(names(NASA), refcols))) :
#>  undefined columns selected
```

Seems like failing earlier (before the GET request) if allowed vars is known ahead of time would be best

- as a possible thing for later, at least with the region fxn, seems they can take a long time to download, maybe caching them would be helpful to users?
- with `enddate = Sys.Date()` is there a possibility this will fail if it's a date that the data isn't ready yet? is there at least a good failure mode when that happens?
