### 2017-12-20 034ab88
Add internal check for user-entered vars
Stop if user enters an invalid `var` for download
