### 2017-12-20 2f4f699
Create .get_NASA function
Both `get_cell` and `get_region` shared the same code, moved to internal_functions.R in `get_NASA`.

Split `GET()` and `content()` with a `stop_for_status()` for user
