### 2017-12-20 a302a20
Add error handling and update documentation regarding endate
Add detail that data often lags for days to a month or more

Check before returning object and stop with message if user requests data that are not available, i.e., POWER website returns a data frame of NA values
