+++ man/get_cell.Rd
@@ -10,7 +10,7 @@ get_cell(lonlat = NULL, vars = c("T2M", "T2MN", "T2MX", "RH2M", "toa_dwn",
 }
 \arguments{
 \item{lonlat}{A length-2 numeric vector giving the decimal degree longitude
-and latitude in that order for cell data to download}
+and latitude in that order for cell data to download.}
 
 \item{vars}{Weather variables to download, defaults to all available.
 Valid variables are:
@@ -35,9 +35,12 @@ the Earth (degrees C)}
 (m/s)} }}
 
 \item{stdate}{Starting date for download, defaults to 01/01/1983 (there is no
-earlier data)}
+earlier data).}
 
-\item{endate}{End date for download, defaults to current date}
+\item{endate}{End date for download, defaults to current date. Note that data
+is often available only with a lag of days to a month or more. The last
+available data will be silently returned with \code{NA} for any values not
+yet reported.}
 }
 \value{
 A tidy \code{\link[base]{data.frame}} object of the requested variable(s)
