### 2017-12-20 c76fa17
Move duplicate functionality to internal_functions.R and write test
