+++ tests/testthat/test-internal_functions.R
@@ -11,6 +11,8 @@ test_that(".check_response proceeds if server is responding", {
   expect_warning(.check_response(url), regexp = NA)
 })
 
+# Check that .check_lonlat* handles incorrect values properly  -----------------
+context(".check_lonlat_cell")
 test_that(".check_lonlat_cell properly reports errors", {
   # out-of-scope latitude
   lonlat <- c(-27.5, 151.5)
@@ -25,6 +27,7 @@ test_that(".check_lonlat_cell properly reports errors", {
   expect_error(.check_lonlat_cell(lonlat))
 })
 
+context(".check_lonlat_region")
 test_that(".check_lonlat_region properly reports errors", {
   # out-of-scope latitude
   lonlat <- c(-181, 181, -90, 90)
@@ -49,3 +52,21 @@ test_that(".check_lonlat_region properly reports errors", {
   lonlat <- c(-180, 180, 90, -90)
   expect_error(.check_lonlat_region(lonlat))
 })
+
+# Check that .create_nasa_df handles data/no data properly ---------------------
+context(".create_nasa_df")
+test_that(".create_nasa_df alerts user that no data are available", {
+  load(system.file("extdata", "NASA_no_data.rda", package = "nasapower"))
+  stdate = as.Date("2017-12-19")
+  endate = as.Date("2017-12-21")
+  expect_error(.create_nasa_df(NASA, stdate, endate))
+})
+
+test_that(".create_nasa_df creates a data frame of data", {
+  load(system.file("extdata", "NASA_with_data.rda", package = "nasapower"))
+  stdate = as.Date("2017-11-1")
+  endate = as.Date("2017-11-3")
+  test <- .create_nasa_df(NASA, stdate, endate)
+  expect_is(test, "data.frame")
+})
+
