### 2017-12-20 f4096eb
Back to 100% test coverage
