### 2017-12-20 efb711c
fix .check_vars test for invalid var
