### 2017-12-20 51211f4
Test for error message in .check_vars
