### 2017-12-20 2031507
Rejig internal functions and tests
