### 2017-12-20 4ed87c8
Set up function to validate user-entered dates
