### 2017-12-20 ea98fc8
Start implementing new date checking function
