### 2017-12-20 c8830d8
Use lubridate to validate user entered date values
