+++ R/get_cell.R
@@ -80,13 +80,14 @@ get_cell <-
            ),
            stdate = "1983-1-1",
            endate = Sys.Date()) {
+
+    #check user inputs, see internal_functions.R for these
     .check_lonlat_cell(lonlat)
-
     .check_vars(vars)
-
     dates <- .check_dates(stdate, endate)
-    stdate <- dates[1]
-    endate <- dates[2]
+
+    stdate <- as.Date(dates[[1]])
+    endate <- as.Date(dates[[2]])
 
     # check if website is responding
     url <-
