### 2017-12-20 1f9387d
Clean up use-case vignette YAML
