### 2017-12-20 9853ab3
Fix bug where reversed dates were not corrected
