### 2017-12-20 200cdeb
Remove extra tests
After moving functions from `get_cell` and `get_region` to internal_functions.R, no need to test for dates out of range in both of these functions. Tested in one spot now for `.check_dates()` function
