### 2017-12-20 #155 @adamhsparks
@sckott, can you add some detail to your first point, "unless you have some edge case stuff, would it not be easier to construct a named list of query params to pass to GET"?

I don't understand what you're suggesting here.

Also, do you have any ideas for caching the data with the region fxn? I'm trying to figure something out, I got nothin. Does _httr_ offer any methods for caching data or capturing data from a failed download?

I've covered all of your other points listed here already.

*check user-entered vars*
https://github.com/adamhsparks/nasapower/commit/034ab8808d117cd303f28136a38910851005cd46

*split `GET` and `content`*
https://github.com/adamhsparks/nasapower/commit/2f4f699e7befe1180ef143e99ef4b141116cdca2

*handle `endate` values*
https://github.com/adamhsparks/nasapower/commit/a302a2089b2b448588623d13e081b853bb341759

*I've also introduced a check and reformat for user-entered date values*
https://github.com/adamhsparks/nasapower/commit/4ed87c8373269dedfb1f0c4526e2bf7e2e899e0a
