### 2017-12-20 adb5974
Reformat the default starting date in documentation
It was 1/1/1983 in documentation. Reformat to follow the standard for `stdate` and `endate` in this package, 1983-01-01
