+++ vignettes/nasapower.Rmd
@@ -65,7 +65,7 @@ Valid `vars` include:
 
     * WS10M - Wind speed at 10m above the surface of the Earth (m/s)
 
-- `stdate` Starting date for download, defaults to 01/01/1983 (there is no
+- `stdate` Starting date for download, defaults to 1983-01-01 (there is no
 earlier data)
 
 - `endate` End date for download, defaults to current date
