### 2017-12-20 #155 @sckott
e.g. instead of this pattern https://github.com/adamhsparks/nasapower/blob/master/R/get_cell.R#L111-L135 can do instead `list(lon = longlat[1], lat = lonlat[2], ...)`, etc. and pass to `GET` as first param just the base URL, and pass that list to `query`

No, there's nothing built in for caching. There's lots of potential solutions for caching. A pattern I use is https://cran.rstudio.com/web/packages/hoardr/ which I think you use in a few of your pkgs?  But there are other solutions: using rappdirs by itself, etc. I think the important part is whether or not you think it's needed (you know better than I), then at least make sure to use the right paths on the users machine (via rappdirs), and make sure to remove any files from a failed download so that you don't end up reading in later a cached file that doesn't have what you expect to have in it. Also, CRAN doesn't like it when you write to the disk in tests/examples that run on on cran checks (hoardr makes this easier by allowing to write to a tempdir, which they are okay with)

what do you mean by "capturing data from a failed download"
