### 2017-12-20 #155 @adamhsparks
Ah, sorry about the caching. I totally understand. My mind had just gone somewhere else with trying to capture data being downloaded, my example of downloading the entire grid for one day takes a while, and thinking if the download was interrupted how to capture that so as to not have to start over again. Might not be possible.

Sure, I can integrate _hoardr_ into this package, easily done.

I've glanced at the httr documentation. I think I understand the `list`, `GET` and `query` structure.

Thanks for the suggestions and clarification.
