### 2017-12-21 3d9dd18
Use query function from httr and fix bugs related to date handling internally
