### 2017-12-21 6f8aac8
create new data and tests for checking data frames if empty
