### 2017-12-21 #155 @adamhsparks
@sckott, I've implemented the use of `query`. I can't seem to get it to allow for multiple values with the same name though. All of the weather parameters are the same key, "p". I can create a list with all ten parameters listed, but `query` only seems to return the first and drops the rest. I've worked around it with a bit cleaner bit of code that still involves `paste` than what I had used. Overall, it does make the code cleaner and easier to follow. Thank you for that suggestion. I now know more about interacting with web APIs.

I've been thinking about the use of _hoardr_ or _rappdirs_ for this package. I don't know if there is much merit to it, but could be convinced otherwise. From my thinking, it's not hard to just use `write.csv` or `fwrite` or `write_csv` or ... and save the data since it's already in a tidy data frame. I was really aiming for this one to be a simple interface to just download and format the data, after that it's up to the user.

Some difficulties, if the user makes multiple requests, e.g., one for Queensland and another for Australia, then you could save time not querying Qld, since it's nested in Australia. However, it would require a fair bit of work with some spatial packages to check what's present and what's requested and getting them to align since you can request pretty much any combination of lat/lon values from the API and it will give you the data for those values.
