### 2017-12-21 af8d4b6
Remove Jorrel as ctb
Original code contributed in early versions no longer part of this package
