### 2017-12-21 1d3591b
Add reference to agroclimatology methods
