+++ inst/paper/paper.bib
@@ -1,3 +1,23 @@
+%% This BibTeX bibliography file was created using BibDesk.
+%% http://bibdesk.sourceforge.net/
+
+%% Created for Adam Sparks at 2017-12-21 15:58:36 +1000 
+
+
+%% Saved with string encoding Unicode (UTF-8) 
+
+
+
+@webpage{stackhouse2017,
+	Author = {Paul W. {Stackhouse, Jr} and David Westberg and William S. Chandler and and Taiping Zhang and James M. Hoell},
+	Date-Added = {2017-12-21 05:54:40 +0000},
+	Date-Modified = {2017-12-21 05:58:20 +0000},
+	Lastchecked = {2017-12-21},
+	Month = {05},
+	Title = {Prediction Of Worldwide Energy Resource (POWER) Agroclimatology Methodology (1.0 Latitude by 1.0 Longitude Spatial Resolution)},
+	Url = {https://power.larc.nasa.gov/documents/Agroclimatology_Methodology.pdf},
+	Urldate = {2017-05-30},
+	Year = {2017}}
 
 @article{bai2010evaluation,
 	Author = {Bai, Jinshun and Chen, Xinping and Dobermann, Achim and Yang, Haishun and Cassman, Kenneth G and Zhang, Fusuo},
