### 2017-12-21 3bfcc35
Edit paper.md
Reword and reorganise for clarity

Add Stackhouse citation.

Clarify what the 1x1 degree is in sq km

Add portion on how to acknowledge NASA-POWER data

Add citation for _APSIM_ package
