+++ inst/paper/paper.md
@@ -17,65 +17,53 @@ tags:
 affiliations:
   index: 1
   name: University of Southern Queensland, Centre for Crop Health, Toowoomba Queensland 4350, Australia
-date: 2017-12-19
+date: 2017-12-21
 ---
 
-# Introduction
+# Summary and Statement of Need
 
-NASA’s POWER (Prediction Of Worldwide Energy Resource) data are
-freely available for download via a
+_nasapower_ [@sparks2017] is an R [@R-base] package providing functionality to
+automate and make NASA-POWER agroclimatology [@stackhouse2017] data retrieval
+faster and easier than using the web interface and also reproducible. Two
+functions are provided, one to query for and return single cell values, the
+other to query and return all cell values in a given region. Both functions
+return the requested data in a tidy data frame [@wickham2014] to allow users to
+focus on using the data rather than importing and tidying. Integrating this data
+retrieval and formatting in R will streamline processes with models such as
+APSIM [@keating2003] and EPIRICE [@savary2012] that can be linked to or are
+implemented in R. Extended documentation is provided with examples of visualising
+the data using _ggplot2_ [@wickham2009], converting it to spatial objects using
+_raster_ [@hijmans2017] and for obtaining solar radiation values and generating
+a .met file using the _APSIM_ R package [@fainges2017] to use in the APSIM
+model.
+
+# About POWER Data
+
+NASA’s POWER (Prediction Of Worldwide Energy Resource) agroclimatology data
+[@stackhouse2017] are freely available for download via a
 [web interface](https://power.larc.nasa.gov/cgi-bin/agro.cgi?email=agroclim@larc.nasa.gov)
-at a grid resolution of 1 degree longitude by 1 degree latitude. Funded through
-the NASA Earth Science Directorate Applied Science Program, the data provide
-daily global coverage from 1983 until near present for all parameters except
-precipitation, which is provided for January 1997 to near present with a several
-month delay. Parameters available for download include: i) top-of-atmosphere
-insolation, ii) insolation on horizontal surface, iii) downward longwave
-radiative flux, iv) average air temperature at two meters, v) minimum air
-temperature at two meters, vi) maximum air temperature at two meters, vii)
-relative humidity at two meters, viii) dew point at two meters, ix)
-precipitation (Jan 1997 to near present with several month delay) and x) wind
-speed at ten meters. The data are widely used in agricultural modelling
-[@bai2010evaluation; @vanwart2013; @vanwart2015] for estimating yields, other
-crop simulation exercises [@ojeda2017] and plant disease modelling
-[@savary2012].
-
-# Using _nasapower_
-
-_nasapower_ [@sparks2017] provides two R [@R-base] functions to automate and
-make data retrieval faster, easier and more reproducible than using the web
-interface. The first, `get_cell()`, downloads NASA-POWER agroclimatology data
-for a given one degree latitude by one degree longitude cell for a range of
-days and years and specified solar radiation or meteorological parameters,
-_e.g._, download all data available for the cell containing Toowoomba,
-Queensland, Australia on January 1, 2017.
-
-```r
-nasa <- get_cell(lonlat = c(-27.5, 151.5), stdate = "2017-1-1",
-                 endate = "2017-1-1")
-```
-
-The second function, `get_region()`, performs the same task for a region defined
-by the four corners, minimum longitude, maximum longitude, minimum latitude, and
-maximum latitude and returns the requested values for the one degree cells that
-all within that region, _e.g._, download all available data for Australia on
-January 1, 2017.
-
-```r
- nasa <- get_region(lonlat = c(112.91972, 159.256088, -55.11694, -9.221099),
-                    stdate = "2017-1-1", endate = "2017-1-1")
-```
-
-Both functions will return the requested data in a tidy data frame
[…]
