+++ inst/paper/paper.bib
@@ -1,12 +1,13 @@
-%% This BibTeX bibliography file was created using BibDesk.
-%% http://bibdesk.sourceforge.net/
 
-%% Created for Adam Sparks at 2017-12-21 15:58:36 +1000 
-
-
-%% Saved with string encoding Unicode (UTF-8) 
-
-
+@manual{fainges2017,
+	Author = {Justin Fainges},
+	Date-Added = {2017-12-21 06:17:25 +0000},
+	Date-Modified = {2017-12-21 06:20:15 +0000},
+	Note = {R package version 0.9.2},
+	Organization = {CSIRO},
+	Title = {{APSIM}: General Utility Functions for the 'Agricultural Production Systems Simulator'},
+	Url = {https://CRAN.R-project.org/package=APSIM},
+	Year = {2017}}
 
 @webpage{stackhouse2017,
 	Author = {Paul W. {Stackhouse, Jr} and David Westberg and William S. Chandler and and Taiping Zhang and James M. Hoell},
