### 2017-12-21 48b3c6a
Update DESCRIPTION
