### 2017-12-21 2b7df06
Remove extranous functionality
