### 2017-12-21 #155 @sckott
Right, repeated params are a bit trickier. Glad it works.

Totally up to you on the caching, makes sense to me what you describe to not do the caching.
