### 2017-12-22 f96aff7
Update documentation to be consistent across files
