+++ man/nasapower.Rd
@@ -6,12 +6,13 @@
 \alias{nasapower-package}
 \title{NASA-POWER Agroclimatology Data from R}
 \description{
-\code{nasapower} provides an R interface to download NASA-POWER
-agroclimatology data and create a tidy data frame for individual cells or a
-specified region.  POWER (Prediction Of Worldwide Energy Resource) data are
-freely available for download through a web interface at a resolution of 1
-degree longitude by 1 degree latitude funded through the NASA Earth Science
-Directorate Applied Science Program.
+Download NASA-POWER agroclimatology data and create a tidy data frame for
+individual cells or a user specified region. nasapower provides an R
+interface to the NASA - POWER API.  POWER (Prediction Of Worldwide Energy
+Resource) data are freely available for download with a resolution of 1 arc
+degree longitude by 1 arc degree latitude and are funded through the NASA
+Earth Science Directorate Applied Science Program. For more on the data
+themselves, please see <https://power.larc.nasa.gov/>.
 }
 \note{
 While \code{nasapower} does not redistribute the data in any way, we
