+++ DESCRIPTION
@@ -8,11 +8,12 @@ Authors@R: person(given = "Adam H.", family = "Sparks",
 URL: https://github.com/adamhsparks/nasapower
 BugReports: https://github.com/adamhsparks/nasapower/issues
 Description: Download NASA-POWER agroclimatology data and create a tidy data
-    frame for individual cells or a specified region. nasapower is a lightweight
-    interface to the NASA - POWER website for R. POWER (Prediction Of Worldwide
-    Energy Resource) data are freely available for download through a web
-    interface at a resolution of 1 arc degree longitude by 1 arc degree latitude
-    funded through the NASA Earth Science Directorate Applied Science Program.
+    frame for individual cells or a user specified region. nasapower provides an
+    R interface to the NASA - POWER API.  POWER (Prediction Of Worldwide Energy
+    Resource) data are freely available for download with a resolution of 1 arc
+    degree longitude by 1 arc degree latitude and are funded through the NASA
+    Earth Science Directorate Applied Science Program. For more on the data
+    themselves, please see <https://power.larc.nasa.gov/>.
 Depends:
     R (>= 3.2.0)
 License: MIT + file LICENSE
