### 2017-12-22 f13b521
Update package for 0.1.3 release on GitHub
