+++ NEWS.md
@@ -1,5 +1,5 @@
 
-# nasapower 0.1.2.9000
+# nasapower 0.1.3
 
 ### Bug fixes
 
