### 2017-12-22 084273a
deprecated get_region and get_cell in favour of get_power
