### 2017-12-22 80703df
Start adding functionality for get_power()
