### 2017-12-22 bfc7a70
Fix dates handling in internal_fucntions
