### 2017-12-22 13b9402
Update documentation and fix internal date handling
