### 2017-12-22 ab4aedd
replace e.g. with i.e.
