### 2017-12-22 b418ca3
Check lonlat and return identifier based on user input
