+++ R/internal_functions.R
@@ -54,119 +54,123 @@
   return(dates)
 }
 
-# validate cell lon lat user provided values -----------------------------------
-.check_lonlat_cell <- function(lonlat) {
-  if (is.null(lonlat) | length(lonlat) != 2 | !is.numeric(lonlat)) {
-    stop("\nlonlat must be provided in a length-2 numeric vector.\n")
-  }
-  if (lonlat[1] < -180 || lonlat[1] > 180) {
-    stop("\nPlease check your longitude, `",
-         paste0(lonlat[1]),
-         "`, to be sure it is valid.\n")
-  }
-  if (lonlat[2] < -90 || lonlat[2] > 90) {
-    stop(
-      "\nPlease check your latitude, `",
-      paste0(lonlat[2]),
-      "`, value to be sure it is valid.\n"
-    )
-  }
-}
-
-# validate region lon lat user provided values ---------------------------------
-.check_lonlat_region <- function(lonlat) {
-  if (is.null(lonlat) | length(lonlat) != 4 | !is.numeric(lonlat)) {
-    stop("\nlonlat must be provided in a length-2 numeric vector.\n")
-  }
-  if (lonlat[1:2] < -180 || lonlat[1:2] > 180) {
-    stop("\nPlease check your longitude, `",
-         paste0(lonlat[1:2]),
-         "`, to be sure it is valid.\n")
-  }
-  if (lonlat[3:4] < -90 || lonlat[3:4] > 90) {
-    stop(
-      "\nPlease check your latitude, `",
-      paste0(lonlat[3:4]),
-      "`, value to be sure it is valid.\n"
-    )
-  }
-  if (lonlat[1] > lonlat[2]) {
-    stop("The first lon value must be the minimum value requested.")
-  }
-  if (lonlat[3] > lonlat[4]) {
-    stop("The first lat value must be the minimum value requested.")
-  }
-}
-
-# validate user entered variables, fail internally if not valid ----------------
-.check_vars <- function(vars, type) {
-  if (all(
-    vars %in% c(
-      "T2M",
-      "T2MN",
-      "T2MX",
-      "RH2M",
-      "toa_dwn",
-      "swv_dwn",
-      "lwv_dwn",
-      "DFP2M",
-      "RAIN",
-      "WS10M"
-    )
-  )) {
-    vars <- vars
-  } else {
-    stop("You have entered an invalid value for `vars`.")
-  }
-
-  # check global vars
-  if (vars > 3 && type == "global") {
-    stop("A max of three vars is allowed for a global query.")
-  }
-}
-
-# check if POWER website is responding -----------------------------------------
-#' @noRd
-.check_response <- function(url) {
-  tryCatch(
-    httr::http_status(httr::GET(url)),
-    error = function(c) {
-      c$message <-
-        paste0("\nThe POWER website does not appear to be responding.\n",
-               "Please try again later.\n")
-      stop(c)
+# validate cell lon lat user provided values and return identifier -------------
+.check_lonlat <- function(lonlat) {
+  if (all.equal(length(lonlat), 2) & is.numeric(lonlat)) {
+    if (lonlat[1] < -180 || lonlat[1] > 180) {
+      stop("\nPlease check your longitude, `",
+           paste0(lonlat[1]),
+           "`, to be sure it is valid.\n")
     }
-  )
+    if (lonlat[2] < -90 || lonlat[2] > 90) {
+      stop(
+        "\nPlease check your latitude, `",
+        paste0(lonlat[2]),
+        "`, value to be sure it is valid.\n"
+      )
+    }
+    identifier <- "SinglePoint"
[…]
