### 2018-01-06 70d8551
Convert to new POWER API
- Start using **crul** in place of **httr**

- Add `parameters` data frame, which lists available parameters with metadata about them

- Provide access to all three communities, not just AG

-
