+++ tests/testthat/test-internal_functions.R
@@ -1,118 +1,49 @@
-context(".check_response")
-# check that .check_response handles web-site responses properly  --------------
 
-test_that(".check_response stops if server not responding", {
-  url <- "http://badurl.gov.au"
-  expect_error(.check_response(url))
-})
+# test that `check_lonlat` handles incorrect values properly  -----------------
+context("check_lonlat")
+test_that("check_lonlat_cell properly reports errors", {
+  # set up pars argument for testing
+  pars <- "T2M"
 
-test_that(".check_response proceeds if server is responding", {
-  url <- "https://www.google.com/"
-  expect_warning(.check_response(url), regexp = NA)
-})
-
-# check that .check_lonlat* handles incorrect values properly  -----------------
-context(".check_lonlat_cell")
-test_that(".check_lonlat_cell properly reports errors", {
-  # out-of-scope latitude
+  # out-of-scope latitude for singlePoint
   lonlat <- c(-27.5, 151.5)
-  expect_error(.check_lonlat_cell(lonlat))
+  expect_error(check_lonlat(lonlat, pars))
 
-  # out-of-scope longitude
+  # out-of-scope longitude for singlePoint
   lonlat <- c(181, 0)
-  expect_error(.check_lonlat_cell(lonlat))
+  expect_error(check_lonlat(lonlat, pars))
 
-  # non-numeric values
+  # non-numeric values for singlePoint
   lonlat <- c(181, "x")
-  expect_error(.check_lonlat_cell(lonlat))
-})
+  expect_error(check_lonlat(lonlat, pars))
 
-context(".check_lonlat_region")
-test_that(".check_lonlat_region properly reports errors", {
-  # out-of-scope latitude
+  # out-of-scope latitude for regional
   lonlat <- c(-181, 181, -90, 90)
-  expect_error(.check_lonlat_region(lonlat))
+  expect_error(check_lonlat(lonlat, pars))
 
-  # out-of-scope longitude
+  # out-of-scope longitude for regional
   lonlat <- c(-180, 180, -91, 91)
-  expect_error(.check_lonlat_region(lonlat))
+  expect_error(check_lonlat(lonlat, pars))
 
-  # incorrect orders
+  # incorrect orders for regional
   lonlat <- c(-180, 180, -91, 91)
-  expect_error(.check_lonlat_region(lonlat))
+  expect_error(check_lonlat(lonlat, pars))
 
-  # non-numeric values
+  # non-numeric values for regional
   lonlat <- c(112.91972, 159.256088, -55.11694, "x")
-  expect_error(.check_lonlat_region(lonlat))
+  expect_error(check_lonlat(lonlat, pars))
 
-  # incorrect order of values requested
+  # incorrect order of values requested for regional
   lonlat <- c(180, -180, -90, 90)
-  expect_error(.check_lonlat_region(lonlat))
+  expect_error(check_lonlat(lonlat, pars))
 
   lonlat <- c(-180, 180, 90, -90)
-  expect_error(.check_lonlat_region(lonlat))
+  expect_error(check_lonlat(lonlat, pars))
 })
 
-# check that .create_nasa_df handles data/no data properly ---------------------
-context(".check_nasa_df")
-test_that(".check_nasa_df alerts user that no data are available", {
-  load(system.file("extdata", "NASA_no_data.rda", package = "nasapower"))
-  expect_error(.check_nasa_df(NASA))
-})
-
-test_that(".check_nasa_df creates a data frame of data", {
-  load(system.file("extdata", "NASA_with_data.rda", package = "nasapower"))
-  expect_error(.check_nasa_df(NASA), NA)
[…]
