### 2018-01-30 #155 @sckott
@adamhsparks still working on changes?
