### 2018-01-30 #155 @adamhsparks
Yeah, I've been sidetracked a bit. I don't know when the new API goes live and the last time I had the chance to test it the site was down.

I do know that  the current version of the package does work, I've had feedback from someone using it with another one of my R packages that models rice diseases. They're steps ahead of me, I've not even done that yet.

I'll try to devote some time to it later this week.
