### 2018-01-31 2b9b3b8
Fix bug in get_cell and get_region
https://github.com/adamhsparks/nasapower/issues/10
