### 2018-01-31 3a4305d
Add Aji as ctb in DESCRIPTION
