+++ DESCRIPTION
@@ -2,9 +2,12 @@ Package: nasapower
 Type: Package
 Title: NASA-POWER Agroclimatology Data from R
 Version: 0.1.3
-Authors@R: person(given = "Adam H.", family = "Sparks",
+Authors@R: c(person(given = "Adam H.", family = "Sparks",
     role = c("aut", "cre"), email = "adamhsparks@gmail.com",
-    comment = c(ORCID = "0000-0002-0061-8359"))
+    comment = c(ORCID = "0000-0002-0061-8359")),
+    person(given = "Aji", family = "Sukarta", role = "ctb",
+    email = "aji.ixb@gmail.com",
+    comment = "Reported bug related to dates"))
 URL: https://github.com/adamhsparks/nasapower
 BugReports: https://github.com/adamhsparks/nasapower/issues
 Description: Download NASA-POWER agroclimatology data and create a tidy data
