### 2018-02-01 5d34702
Update NEWS.md for v0.1.4 release
