### 2018-02-02 b809afa
Fix incorrect call to get_power internally, should be power_query
