### 2018-02-02 99dd072
Correct dates formatting for querying NASA/POWER server
