### 2018-02-03 8803ea7
Read CSV result from POWER server
