### 2018-02-03 #155 @adamhsparks
Hi @sckott, I know you've seen that I fixed a bug in the master branch and updated the currently released version this week.

I've also spent a few minutes on the devel version working with the new API. I've successfully run a test and retrieved data using the new API with `crul`! I need some time to tidy up and write tests since this is basically a complete rewrite of the package, but I'm happy to say that I'm making progress, albeit slowly.
