### 2018-02-03 ddcb693
Add jsonlite as dependency
