### 2018-02-13 #155 @adamhsparks
Just to update you, @sckott, I'm short on time right now to work on this. However, any time I've tried to query multiple vars it fails telling me I have an incorrect var specified. However, if I query any of the same vars singly, it succeeds for all of them.

I need some more time to really focus on this, but I likely won't have that until sometime next month. Is that OK?
