### 2018-03-30 9aa1a60
Update static docs
