### 2018-03-30 3922319
Fix parameter checks
If any are NOT in the parameters list, stop as invalid
