### 2018-03-30 ac96adc
Fix bug where POWER reported invalid parameters
For some reason the POWER API doesn't parse a unicode comma properly. Wrapping the parameters in `I()` allows the pars to pass through properly formatted. Will need to explore and make sure that there are not other parts of the query that require this as well
