+++ R/internal_functions.R
@@ -223,7 +223,7 @@ power_query <- function(community,
     query_list <- list(
       request = "execute",
       identifier = lonlat_identifier$identifier,
-      parameters = pars,
+      parameters = I(pars),
       startDate = dates[[1]],
       endDate = dates[[2]],
       userCommunity = community,
@@ -239,7 +239,7 @@ power_query <- function(community,
     query_list <- list(
       request = "execute",
       identifier = lonlat_identifier$identifier,
-      parameters = pars,
+      parameters = I(pars),
       startDate = dates[[1]],
       startDate = dates[[2]],
       userCommunity = community,
@@ -254,7 +254,7 @@ power_query <- function(community,
     query_list <- list(
       request = "execute",
       identifier = lonlat_identifier$identifier,
-      parameters = pars,
+      parameters = I(pars),
       startDate = dates[[1]],
       startDate = dates[[2]],
       userCommunity = community,
