### 2018-03-30 7107e93
Suppress messages about coltypes from readr
