### 2018-03-30 67310e0
Fix bug checking lonlat identifier
