### 2018-03-30 1ad2d3b
Update documentation for get_power
