+++ man/get_power.Rd
@@ -85,6 +85,18 @@ for a given month, averaged for that month over the 30-year period (Jan.
 1984 - Dec. 2013).} }
 }
 
+\examples{
+# Fetch temperature and relative humidity for January 1 1985
+
+\dontrun{
+power <- get_power(community = "AG",
+                  lonlat = c(-179.5, -89.5),
+                  pars = c("RH2M", "T2M"),
+                  dates = "1985-01-01",
+                  temporal_average = "daily")
+}
+
+}
 \references{
 \url{https://power.larc.nasa.gov/documents/Agroclimatology_Methodology.pdf}
 }
