### 2018-03-30 1db4aee
Use tryCatch in reading results from server
