### 2018-03-31 3bc5bc3
Update internal functions to use tryCatch more efficiently
