### 2018-03-31 bfe9310
Update use-case vignette for new POWER data format
