### 2018-03-31 aecb644
prettify message for bbox
