### 2018-03-31 c895b7c
Update bbox documentation
