### 2018-03-31 3ca2b57
Fix issues with bbox
Check values to make sure that we're not requesting too many cells. 100 max!

Wrap `bbox` par in `I()` for query
