### 2018-03-31 3bfb197
Update message for error on bbox too large
