### 2018-03-31 0aad475
Update README and vignette
Update with information for new, updated data, communities and functions in _nasapower_
