+++ README.Rmd
@@ -46,66 +46,45 @@ devtools::install_github("adamhsparks/nasapower", build_vignettes = TRUE)
 _nasapower_ aims to make it quick, easy and efficient to automate downloading
 NASA-POWER agroclimatology data in your R session as a tidy data frame.
 
-_nasapower_ only provides functions, `get_cell()`, for a given 1˚ longitude by
-1˚ latitude cell and `get_region()`, both of which will which will download
+_nasapower_ only provides one function, `get_power()`, which will download
 specified variables and return a tidy data frame of the requested data. Weather
-variables can be specified by using the `vars` argument. 
+variables can be specified by using the `pars` argument. 
 
-Both functions accept four arguments.
+## Using get_power()
 
-```{r arguments, eval=FALSE}
-get_cell(lonlat = NULL, vars = c("T2M", "T2MN", "T2MX", "RH2M"),
-  stdate = "1983-1-1", endate = Sys.Date())
+The `get_power()` function requires five arguments as seen in this example,
+which will fetch relative humidity (RH2M) and temperature (T2M) for the year
+1985 on a daily time-step.
+
+```{r example, eval=FALSE}
+power <- get_power(community = "AG",
+                   lonlat = c(-179.5, -89.5),
+                   pars = c("RH2M", "T2M"),
+                   dates = c("1985-01-01", "1985-12-31"),
+                   temporal_average = "daily")
 ```
 
-- `lonlat` for `get_cell()` A length-2 numeric vector giving the decimal degree
-longitude and and latitude in that order for cell data to download
+The arguments are:
 
-- `lonlat` for `get_region()` A length-4 numeric vector of the minimum
-longitude, maximum longitude, minimum latitude and maximum latitude.
+- `community`, a text string with valid values of: "AG" (Agroclimatology), "SSE"
+(Surface meteorology and Solar Energy) or "SB" (Sustainable Buildings). The
+selected user community will affect the units of the parameter and the temporal
+display of time series data (_e.g._ Agroclimatology will use MJ/m^2^/day for
+radiation units, while SSE and SB use kW/m^2^/day as units).
 
-- `vars` Weather variables to query for download, defaults to all available.
-Valid `vars` include:
+- `lonlat`, a length-2 numeric vector giving the decimal degree longitude and
+and latitude coordinates in that order for cell data to download or a length-4
+numeric vector giving the decimal degree longitude and and latitude coordinates
+forming a bounding box as ymin, xmin, ymax, xmax in that order.
 
-    * toa_dwn - Average top-of-atmosphere insolation (MJ/m^2/day)
+- `pars`, a character vector of weather variables to query for download. For a
+complete listing of valid `pars`, please see column 1 of the package included
+data, `parameters`, e.g., using RStudio, `View(parameters)`.
 
-    * swv_dwn - Average insolation incident on a horizontal surface (MJ/m^2/day)
+- `dates`, a vector of start and end dates for which to query the NASA-POWER API
 
-    * lwv_dwn - Average downward longwave radiative flux (MJ/m^2/day)
-
-    * T2M - Average air temperature at 2m above the surface of the Earth
-    (degrees C)
-
-    * T2MN - Minimum air temperature at 2m above the surface of the Earth
-    (degrees C)
-
-    * T2MX - Maximum air temperature at 2m above the surface of the Earth
-    (degrees C)
-
-    * RH2M - Relative humidity at 2m above the surface of the Earth (%)
-
-    * DFP2M - Dew/Frost point temperature at 2m above the surface of the Earth
-    (degrees C)
-
-    * RAIN - Average precipitation (mm/day)
-
[…]
