### 2018-03-31 0ed269c
Fix degrees in README
