+++ README.Rmd
@@ -20,8 +20,8 @@ it quick and easy to automate downloading
 [NASA-POWER](https://power.larc.nasa.gov) agroclimatology data in your R session
 as a tidy data frame for agricultural analysis and use in modelling or other
 purposes. POWER (Prediction Of Worldwide Energy Resource) data are freely
-available for download through a web interface at a resolution of 1˚ longitude
-by 1˚ latitude.
+available for download through a web interface at a resolution of 0.5˚ longitude
+by 0.5˚ latitude.
 
 Please see
 <https://power.larc.nasa.gov/cgi-bin/agro.cgi?email=agroclim@larc.nasa.gov>
