### 2018-03-31 34f628f
Check temporal_average for pars
