+++ DESCRIPTION
@@ -24,6 +24,7 @@ Imports:
     jsonlite,
     readr,
     stringi,
+    tools,
     utils
 RoxygenNote: 6.0.1
 Roxygen: list(markdown = TRUE)
