### 2018-03-31 e29bb5e
Add message for checking temporal_average
