### 2018-03-31 b5d7ce7
Tidy up code formatting
