### 2018-03-31 574485f
Condense functionality
Move `check_tempavg` functionality into `check_pars`. Less steps in `get_power()` and integrates checking since the pars and temporal average are linked and are checked together in this function.
