### 2018-03-31 63edbdc
Update static docs
