### 2018-03-31 0d937d2
Reorder checks
Check pars first then check lonlat since the number of pars specified will determine if global can be used
