+++ R/get_power.R
@@ -91,8 +91,8 @@ get_power <- function(community = NULL,
   # user input checks and formatting -------------------------------------------
   # see internal_functions.R for these functions
   dates <- check_dates(dates)
-  lonlat_identifier <- check_lonlat(lonlat, pars)
   pars <- check_pars(pars)
+  lonlat_identifier <- check_lonlat(lonlat, pars)
   community <- check_community(community)
 
   # submit query ---------------------------------------------------------------
