### 2018-04-10 ab172a2
grammar and spaces in vignette
