### 2018-04-10 5fa7d2e
fix issues with temporal_average errors
