### 2018-05-19 b045523
Fix up deprecated functions
