### 2018-05-19 02faab5
Add tests for deprecated functions
