### 2018-05-19 f11ab37
Undo changes to deprecated functions
