### 2018-05-19 3ffe159
Remove extra line of code in Internal_functions.R
