### 2018-05-19 f08f80d
Correct tests for deprecated functions
