### 2018-05-19 040cd04
Add tests for date handling and checking
