### 2018-05-19 41e894e
Split tests for dates and latlon into own files
