### 2018-05-19 49e7542
Add lettercase import
