### 2018-05-19 050ed5f
Clean up internal functions
- use `lettercase` to convert to titlecase format for `temporal_average`

- Clean up message, no need to use `paste()`

- Add nocov where necessary
