### 2018-05-19 ba3a62e
add tests for check_pars
