### 2018-05-19 1aa7cdb
add nocov to return() statement
