### 2018-05-20 505eda4
stop if latlon not provided
