### 2018-05-20 31f9a9d
edit message for lonlat stop
