### 2018-05-20 96d552f
Update power_url
