### 2018-05-20 1103287
Test that lonlat is provided
