### 2018-05-20 08e964e
Add test for single point and too large bbox
