### 2018-05-20 a7e8b7b
some cleanup in vignettes
