### 2018-05-20 1237048
Prettify messages and clean up internal_functions.R
