### 2018-05-20 31623cd
Add test to check validity of latitudes in bbox
