### 2018-05-20 #155 @adamhsparks
Hi @sckott, I'm still working on this. I had some time this weekend to devote to it and have been writing tests. I'm nearly 1/2 done with the tests and then need to rewrite the vignettes.

Sorry for the delays. "Real" work keeps getting in the way. I've also had to make some decisions about how/what data to access. I can't find a good way to reliably access the ESRI REST API for raster data through R yet. So I will just be focusing on the regular single point/regional API that Power provides, for now with hopes that in the future I can sort the access to raster data directly.
