+++ man/get_power.Rd
@@ -12,8 +12,7 @@ get_power(community = NULL, lonlat = NULL, pars = NULL, dates = NULL,
 argument details for more.}
 
 \item{lonlat}{A numeric vector of geographic coordinates for a cell or region
-entered as x, y coordinates.  Leave unspecified for global coverage.  See
-argument details for more.}
+entered as x, y coordinates.  See argument details for more.}
 
 \item{pars}{A character vector of solar or meteorological variables to
 download.  See the \code{Value} field of \code{\link{parameters}} for a full list of valid
@@ -57,15 +56,13 @@ powered renewable energy systems.} }
  \describe{ \item{For a single
 point}{To get a specific cell, 1/2 x 1/2 degree, supply a length-2 numeric
 vector giving the decimal degree longitude and latitude in that order for
-data to download, \emph{e.g.}, \code{lonlat = c(-179.5, -89.5)}.}
+data to download, \emph{e.g.}, \code{lonlat = c(-179.5, -89.5)}.}}
 
 \item{For regional coverage}{To get a region, supply a length-4 numeric
 vector as \code{lonlat = c(ymin, xmin, ymax, xmax)} in that order for a given
 region, \emph{e.g.}, a bounding box for the southwestern corner of Australia:
 \code{lonlat = c(112.5, 122.5, -55.5, -45.5)}. \emph{Max bounding box is 10 x 10
 degrees} of 1/2 x 1/2 degree data, \emph{i.e.}, 100 points maximum in total.}
-
-\item{For global coverage}{Not used.} }
 }
 
 \section{Argument details for \code{dates}}{
