### 2018-05-20 38c88d9
Correct bbox lonlat order in internal_functions
