### 2018-05-20 52015e0
Add linebreaks to messages
