### 2018-05-20 4e34ba7
Add tests for lonlat
