### 2018-05-20 d753331
Correct lonlat checks
