### 2018-05-20 2f32854
add lonlat tests
