### 2018-05-20 6f794b8
Correct documentation issues
