+++ man/get_power.Rd
@@ -40,7 +40,8 @@ respective sections following below.
 must be supplied. This  will affect the units of the parameter and the
 temporal display of time series data.
 
-\describe{ \item{AG}{Provides access to the Agroclimatology Archive, which
+\describe{
+\item{AG}{Provides access to the Agroclimatology Archive, which
 contains industry-friendly parameters formatted for input to crop models.}
 
 \item{SB}{Provides access to the Sustainable Buildings Archive, which
@@ -49,20 +50,23 @@ parameters in multi-year monthly averages.}
 
 \item{SSE}{Provides access to the Renewable Energy Archive, which contains
 parameters specifically tailored to assist in the design of solar and wind
-powered renewable energy systems.} }
+powered renewable energy systems.}
+}
 }
 
 \section{Argument details for \code{lonlat}}{
- \describe{ \item{For a single
-point}{To get a specific cell, 1/2 x 1/2 degree, supply a length-2 numeric
-vector giving the decimal degree longitude and latitude in that order for
-data to download, \emph{e.g.}, \code{lonlat = c(-179.5, -89.5)}.}}
+
+\describe{
+\item{For a single point}{To get a specific cell, 1/2 x 1/2 degree, supply a
+length-2 numeric vector giving the decimal degree longitude and latitude in
+that order for data to download, \emph{e.g.}, \code{lonlat = c(-179.5, -89.5)}.}
 
 \item{For regional coverage}{To get a region, supply a length-4 numeric
 vector as \code{lonlat = c(xmin, xmax, ymin, ymax)} in that order for a given
 region, \emph{e.g.}, a bounding box for the southwestern corner of Australia:
 \code{lonlat = c(112.5, 122.5, -55.5, -45.5)}. \emph{Max bounding box is 10 x 10
 degrees} of 1/2 x 1/2 degree data, \emph{i.e.}, 100 points maximum in total.}
+}
 }
 
 \section{Argument details for \code{dates}}{
