### 2018-05-20 0faf77e
Move bbox checks to same section of file
