+++ tests/testthat/test-check_latlon.R
@@ -71,21 +71,6 @@ test_that("check_lonlat handles single point properly", {
   expect_equal(test$identifier, "SinglePoint")
 })
 
-test_that("check_lonlat handles bboxes that are too large", {
-  expect_error(check_lonlat(lonlat = c(-179.5, 179.5, -89.5, 89.5), pars),
-                 regexp = "Please provide correct bounding box values*")
-})
-
-test_that("check_lonlat checks order of the latitude values", {
-  expect_error(check_lonlat(lonlat = c(-179.5, 179.5, 89.5, -89.5), pars),
-               regexp = "\nThe first `lat` value must be the minimum value.\n")
-})
-
-test_that("check_lonlat checks order of the lat values", {
-  expect_error(check_lonlat(lonlat = c(179.5, -179.5, -89.5, 89.5), pars),
-               regexp = "\nThe first `lon` value must be the minimum value.\n")
-})
-
 test_that("check_lonlat checks validity of single lat values", {
   expect_error(check_lonlat(lonlat = c(179.5, 91), pars),
                regexp = "Please check your latitude, `91`, value to be sure it is valid.\n")
@@ -97,6 +82,21 @@ test_that("check_lonlat checks validity of single lon values", {
 })
 
 # bbox checks ------------------------------------------------------------------
+
+test_that("check_lonlat handles bboxes that are too large", {
+  expect_error(check_lonlat(lonlat = c(-179.5, 179.5, -89.5, 89.5), pars),
+               regexp = "Please provide correct bounding box values*")
+})
+
+test_that("check_lonlat checks order of the latitude values", {
+  expect_error(check_lonlat(lonlat = c(-179.5, 179.5, 89.5, -89.5), pars),
+               regexp = "\nThe first `lat` value must be the minimum value.\n")
+})
+
+test_that("check_lonlat checks order of the lat values", {
+  expect_error(check_lonlat(lonlat = c(179.5, -179.5, -89.5, 89.5), pars),
+               regexp = "\nThe first `lon` value must be the minimum value.\n")
+})
 
 test_that("check_lonlat checks validity of bbox latmin values", {
   expect_error(check_lonlat(lonlat = c(-179.5, -179.5, 91, 90), pars),
