### 2018-05-20 d439963
Start preparing power_query test
