### 2018-05-20 8ea5fd7
Fix bugs in internal_functions.R
