### 2018-05-20 5acb098
Fix code in vignette example
