### 2018-05-23 a4c9a7c
Fix bug checking parameters and temporal average
