### 2018-05-24 d51932c
Minor fixes for test
Actually read the help file for `covr::context()` it was enlightening
