### 2018-05-24 93c30ea
Add tests for community
