### 2018-05-24 f00a50f
Test more lines in internal_functions.R
