### 2018-05-24 1817e8e
Don't need return() in this function
Short and easy to read it.
