### 2018-05-31 ef53d3a
Correct author field and remove lettercase dependency
