+++ DESCRIPTION
@@ -2,7 +2,7 @@ Package: nasapower
 Type: Package
 Title: NASA-POWER Agroclimatology Data from R
 Version: 1.0.0.9000
-Authors@R: person(given = "Adam H.", family = "Sparks",
+Authors@R: c(person(given = "Adam H.", family = "Sparks",
     role = c("aut", "cre"), email = "adamhsparks@gmail.com",
     comment = c(ORCID = "0000-0002-0061-8359")),
     person(given = "Aji", family = "Sukarta", role = "ctb",
@@ -23,7 +23,6 @@ License: MIT + file LICENSE
 Imports:
     crul,
     httr,
-    lettercase,
     lubridate,
     jsonlite,
     readr,
