### 2018-05-31 ea1b0ef
Update README
Add WIP badge and statement

Correct link to POWER website, old link no longer exists
