+++ README.Rmd
@@ -12,6 +12,8 @@ library(nasapower)
 [![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/adamhsparks/nasapower?branch=master&svg=true)](https://ci.appveyor.com/project/adamhsparks/nasapower)
 [![Coverage Status](https://img.shields.io/codecov/c/github/adamhsparks/nasapower/master.svg)](https://codecov.io/github/adamhsparks/nasapower?branch=master)
 [![DOI](https://zenodo.org/badge/109224461.svg)](https://zenodo.org/badge/latestdoi/109224461)
+[![Project Status: WIP – Initial development is in progress, but there has not yet been a stable, usable release suitable for the public.](http://www.repostatus.org/badges/latest/wip.svg)](http://www.repostatus.org/#wip)
+
 
 ## Introduction
 
@@ -24,10 +26,12 @@ available for download through a web interface at a resolution of 0.5˚ longitud
 by 0.5˚ latitude.
 
 Please see
-<https://power.larc.nasa.gov/cgi-bin/agro.cgi?email=agroclim@larc.nasa.gov>
-for the agroclimatology data web interface.
+<https://power.larc.nasa.gov/> for more on the data and other ways to access it.
 
 ### Quick start
+
+With the changes to the NASA/POWER data and API, this package is a
+work-in-progress under active development, install at your own risk.
 
 _nasapower_ is not available from CRAN, only GitHub. It can easily be installed
 using the following code:
