### 2018-05-31 3f84886
Punctuation in README
