+++ README.Rmd
@@ -113,7 +113,7 @@ Applied Science Program."
 * License: MIT
 
 * Get citation information for `nasapower` in R by typing
-`citation(package = "nasapower")`
+`citation(package = "nasapower")`.
 
 * Please note that this project is released with a
 [Contributor Code of Conduct](CONDUCT.md).
@@ -122,7 +122,7 @@ By participating in this project you agree to abide by its terms.
 * The U.S. Earth System Research Laboratory, Physical Science Division of the
 National Atmospheric & Oceanic Administration (NOAA) maintains a list of
 gridded climate data sets that provide different data and different resolutions
-<https://www.esrl.noaa.gov/psd/data/gridded/>
+<https://www.esrl.noaa.gov/psd/data/gridded/>.
 
 ## References
 <https://power.larc.nasa.gov>
