### 2018-05-31 a1441c5
Update static docs
