+++ docs/articles/global_surface.Rmd
@@ -1,70 +1,0 @@
----
-title: "Using nasapower to Create an Agroclimatology Global Surface"
-author: "Adam H. Sparks"
-date: "`r Sys.Date()`"
-output: rmarkdown::html_vignette
-vignette: >
-  %\VignetteIndexEntry{Using nasapower to Create an Agroclimatology Global Surface}
-  %\VignetteEngine{knitr::rmarkdown}
-  %\VignetteEncoding{UTF-8}
----
-```{r check_packages, echo=FALSE, messages=FALSE, warning=FALSE}
-required <- c("dplyr", "ggplot2", "raster", "virdis")
-
-if (!all(unlist(lapply(required, function(pkg)
-requireNamespace(pkg, quietly = TRUE)))))
-
-knitr::opts_chunk$set(
-eval = FALSE,
-collapse = TRUE,
-comment = "#>",
-fig.width = 7,
-fig.height = 7,
-fig.align = "center"
-)
-```
-
-Creating a complete agroclimatology (>-60˚, <60˚) global surface of NASA -
-POWER data is easy using _nasapower_. However, the process is a bit
-time-intensive.
-
-## Load libraries
-
-```{r libraries, eval=TRUE, message=FALSE}
-library(ggplot2)
-library(nasapower)
-library(viridis)
-```
-
-## Query the server
-
-With the lon/lat values to provide to the `get_nasa()` function, we can now set
-up the data download.
-
-```{r fetch_data, eval=TRUE}
-
-surface <- get_region(lonlat = c(-180, 180, -60, 60), vars = c("T2M", "RH2M"))
-
-```
-
-## Map the data
-
-Using _ggplot2_ it is possible to make simple maps of the data.
-
-```{r}
-# Plot temperature
-ggplot(surface, aes(x = LON, y = LAT, fill = T2M)) +
-  geom_tile() +
-  scale_fill_viridis() +
-  theme_bw() +
-  coord_cartesian() +
-  ggtitle("Average Temperature (˚C)")
-
-# Plot RH
-ggplot(surface, aes(x = LON, y = LAT, fill = RH2M)) +
-  geom_tile() +
-  scale_fill_viridis(option = "C") +
-  theme_bw() +
-  coord_cartesian() +
-  ggtitle("Relative Humidity (%)")
-```
