### 2018-06-02 1cf7852
Clean up documentation, remove deprecated functions
