### 2018-06-02 e1b0d19
Add more information when error occurs in pars
Tell user which `pars` are invalid
