### 2018-06-02 ee02eaa
tests for Single Point
