### 2018-06-02 0dd644e
Remove get_cell and get_region
These are deprecated functions
