### 2018-06-02 dc36019
Fix temporal average bug
Integrate the `temporal_average` and `pars` objects into a list fed to the query.

This makes sure that the `temporal_average` object is converted to upper as POWER requires and reduces the number of arguments in the `power_query` function
