### 2018-06-02 ead56ba
Fix temporal_average toupper check
wrap in a `if !is.null` statement
