### 2018-06-02 91122b4
Update tests for check_pars
Update to test for new functionality where a list is returned not just a vector of characters

Also update where the error message has been customised to be more user friendly
