### 2018-06-02 ced9a28
Update documentation
Add option for "Global" coverage
