+++ tests/testthat/test-check_latlon.R
@@ -1,48 +1,48 @@
 
-# lonlat checks ----------------------------------------------------------------
+# latlon checks ----------------------------------------------------------------
 context("Test that check_latlon function handles lat lon strings correctly")
-test_that("check_lonlat_cell properly reports errors", {
+test_that("check_latlon_cell properly reports errors", {
   # set up pars argument for testing
   pars <- "T2M"
 
   # out-of-scope latitude for singlePoint
-  lonlat <- c(-27.5, 151.5)
-  expect_error(check_lonlat(lonlat, pars))
+  latlon <- c(151.5, -27.5)
+  expect_error(check_latlon(latlon, pars))
 
   # out-of-scope longitude for singlePoint
-  lonlat <- c(181, 0)
-  expect_error(check_lonlat(lonlat, pars))
+  latlon <- c(181, 0)
+  expect_error(check_latlon(latlon, pars))
 
   # non-numeric values for singlePoint
-  lonlat <- c(181, "x")
-  expect_error(check_lonlat(lonlat, pars))
+  latlon <- c(181, "x")
+  expect_error(check_latlon(latlon, pars))
 
   # out-of-scope latitude for regional
-  lonlat <- c(-181, 181, -90, 90)
-  expect_error(check_lonlat(lonlat, pars))
+  latlon <- c(-181, 181, -90, 90)
+  expect_error(check_latlon(latlon, pars))
 
   # out-of-scope longitude for regional
-  lonlat <- c(-180, 180, -91, 91)
-  expect_error(check_lonlat(lonlat, pars))
+  latlon <- c(-180, 180, -91, 91)
+  expect_error(check_latlon(latlon, pars))
 
   # incorrect orders for regional
-  lonlat <- c(-180, 180, -91, 91)
-  expect_error(check_lonlat(lonlat, pars))
+  latlon <- c(-180, 180, -91, 91)
+  expect_error(check_latlon(latlon, pars))
 
   # non-numeric values for regional
-  lonlat <- c(112.91972, 159.256088, -55.11694, "x")
-  expect_error(check_lonlat(lonlat, pars))
+  latlon <- c(112.91972, 159.256088, -55.11694, "x")
+  expect_error(check_latlon(latlon, pars))
 
   # incorrect order of values requested for regional
-  lonlat <- c(180, -180, -90, 90)
-  expect_error(check_lonlat(lonlat, pars))
+  latlon <- c(180, -180, -90, 90)
+  expect_error(check_latlon(latlon, pars))
 
-  lonlat <- c(-180, 180, 90, -90)
-  expect_error(check_lonlat(lonlat, pars))
+  latlon <- c(-180, 180, 90, -90)
+  expect_error(check_latlon(latlon, pars))
 })
 
-test_that("check_lonlat stops if more than three are requested for global", {
-  lonlat <- NULL
+test_that("check_latlon stops if more than three are requested for global", {
+  latlon <- NULL
   pars <-  c(
     "T2M",
     "T2MN",
@@ -55,73 +55,72 @@ test_that("check_lonlat stops if more than three are requested for global", {
     "RAIN",
     "WS10M"
   )
-  expect_error(check_lonlat(lonlat, pars))
+  expect_error(check_latlon(latlon, pars))
 })
 
-test_that("check_lonlat stops if no `lonlat` is supplied", {
-  expect_error(check_lonlat(lonlat = NULL, pars),
-               regexp = "*You must provide a `lonlat` (maximum 100 points total or 10x10 cells)*")
+test_that("check_latlon stops if no `latlon` is supplied", {
+  expect_error(check_latlon(latlon = NULL, pars),
[…]
