### 2018-06-02 e291149
Rejig latlon to match POWER requests
