+++ R/internal_functions.R
@@ -141,41 +141,40 @@ check_pars <-
   }
 
 #' @noRd
-check_lonlat <-
-  function(lonlat, pars) {
+check_latlon <-
+  function(latlon, pars) {
     bbox <- NULL
-    if (is.null(lonlat)) {
+    if (is.null(latlon)) {
       stop(call. = FALSE,
-           "\nYou must provide a `lonlat` (maximum 100 points total or 10x10 cells).\n")
+           "\nYou must provide a `latlon` (maximum 100 points total or 10x10 cells).\n")
     }
-
-    if (length(lonlat) == 2 && is.numeric(lonlat)) {
-      if (lonlat[1] < -180 || lonlat[1] > 180) {
-        stop(
-          call. = FALSE,
-          "\nPlease check your longitude, `",
-          paste0(lonlat[1]),
-          "`, to be sure it is valid.\n"
-        )
-      }
-      if (lonlat[2] < -90 || lonlat[2] > 90) {
+    if (length(latlon) == 2 && is.numeric(latlon)) {
+      if (latlon[1] < -90 || latlon[1] > 90) {
         stop(
           call. = FALSE,
           "\nPlease check your latitude, `",
-          paste0(lonlat[2]),
+          paste0(latlon[1]),
+          "`, to be sure it is valid.\n"
+        )
+      }
+      if (latlon[2] < -180 || latlon[2] > 180) {
+        stop(
+          call. = FALSE,
+          "\nPlease check your longitude, `",
+          paste0(latlon[2]),
           "`, value to be sure it is valid.\n"
         )
       }
       message(
-        "\nFetching single point data for lon ", lonlat[1], ", lat ", lonlat[2],
+        "\nFetching single point data for lat ", latlon[1], ", lon ", latlon[2],
         "\n"
       )
       identifier <- "SinglePoint"
-      lon <- lonlat[1]
-      lat <- lonlat[2]
+      lat <- latlon[1]
+      lon <- latlon[2]
 
-    } else if (length(lonlat) == 4 && is.numeric(lonlat)) {
-      if ((lonlat[[2]] - lonlat[[1]]) * (lonlat[[4]] - lonlat[[3]]) * 4 > 100) {
+    } else if (length(latlon) == 4 && is.numeric(latlon)) {
+      if ((latlon[[3]] - latlon[[1]]) * (latlon[[4]] - latlon[[2]]) * 4 > 100) {
         stop(
           call. = FALSE,
           "\nPlease provide correct bounding box values. The bounding box\n",
@@ -184,66 +183,68 @@ check_lonlat <-
         )
       }
 
-      if (any(lonlat[1] < -180 ||
-              lonlat[2] < -180 ||
-              lonlat[1] > 180 ||
-              lonlat[2] > 180)) {
-        stop(
-          call. = FALSE,
-          "\nPlease check your longitude, `",
-          lonlat[1], "`, `", lonlat[2],
-          "`, values to be sure they are valid.\n"
-        )
-      }
-      if (any(lonlat[3] < -90 ||
-              lonlat[4] < -90 ||
-              lonlat[3] > 90 ||
-              lonlat[4] > 90)) {
+      if (any(latlon[1] < -90 ||
+              latlon[3] < -90 ||
+              latlon[1] > 90 ||
+              latlon[3] > 90)) {
         stop(
           call. = FALSE,
           "\nPlease check your latitude, `",
-          lonlat[3], "`, `", lonlat[4],
+          latlon[1], "`, `", latlon[3],
           "`, values to be sure they are valid.\n"
         )
       }
-      if (lonlat[1] > lonlat[2]) {
-        stop(call. = FALSE,
[…]
