### 2018-06-03 0874179
Fix documentation and add check for Global
call new function to check latlon when == "Global" and set `temporal_average` to "Climatology"
