### 2018-06-03 d455ae9
Update get_power internal checks
The new `check_for_global` returns a list of `latlon` and `temporal_average` values to feed other functions
