### 2018-06-03 9991507
Replace lonlat with latlon in vignettes
