### 2018-06-03 295f590
Fix bug in check_for_global
names in the list object out were not properly assigned
