### 2018-06-03 672da13
Use new list object output from check_for_global
