+++ README.Rmd
@@ -62,7 +62,7 @@ which will fetch relative humidity (RH2M) and temperature (T2M) for the year
 
 ```{r example, eval=FALSE}
 power <- get_power(community = "AG",
-                   lonlat = c(-179.5, -89.5),
+                   lonlat = c(-89.5, -179.5),
                    pars = c("RH2M", "T2M"),
                    dates = c("1985-01-01", "1985-12-31"),
                    temporal_average = "daily")
