### 2018-06-03 5c21304
Replace `lonlat` with `latlon` in README
