### 2018-06-03 eaaf2ac
Correct order of lat lon in a few places and knit README
