+++ README.Rmd
@@ -76,8 +76,8 @@ selected user community will affect the units of the parameter and the temporal
 display of time series data (_e.g._ Agroclimatology will use MJ/m^2^/day for
 radiation units, while SSE and SB use kW/m^2^/day as units).
 
-- `latlon`, a length-2 numeric vector giving the decimal degree longitude and
-and latitude coordinates in that order for cell data to download or a length-4
+- `latlon`, a length-2 numeric vector giving the decimal degree latitude and
+and longitude coordinates in that order for cell data to download or a length-4
 numeric vector giving the decimal degree longitude and and latitude coordinates
 forming a bounding box as ymin, xmin, ymax, xmax in that order.
 
@@ -125,6 +125,7 @@ gridded climate data sets that provide different data and different resolutions
 <https://www.esrl.noaa.gov/psd/data/gridded/>.
 
 ## References
+
 <https://power.larc.nasa.gov>
 
 <https://power.larc.nasa.gov/documents/Agroclimatology_Methodology.pdf>
