+++ vignettes/nasapower.Rmd
@@ -47,8 +47,8 @@ selected user community will affect the units of the parameter and the temporal
 display of time series data (_e.g._ Agroclimatology will use MJ/m^2^/day for
 radiation units, while SSE and SB use kW/m^2^/day as units).
 
-- `latlon`, a length-2 numeric vector giving the decimal degree longitude and
-and latitude coordinates in that order for cell data to download or a length-4
+- `latlon`, a length-2 numeric vector giving the decimal degree latitude and
+and longitude coordinates in that order for cell data to download or a length-4
 numeric vector giving the decimal degree longitude and and latitude coordinates
 forming a bounding box as ymin, xmin, ymax, xmax in that order. When specifying
 a region, the maximum amount of cells that can be downloaded are 100, i.e., a 10
