### 2018-06-03 1269ddc
Correct order for latlon in example region
