### 2018-06-03 023efc9
Fit bounding box example w/in restrictions
