### 2018-06-03 479144e
Remove yml file from top level
