### 2018-06-03 9b14fb3
Fix regional test for get_power()
