### 2018-06-03 3f24274
Update documentation
start/end dates not required for global coverage
