+++ man/get_power.Rd
@@ -22,7 +22,8 @@ the Parameter Dictionary.  If downloading global coverage a maximum of 3
 \code{pars} can be specified at one time.}
 
 \item{dates}{A character vector of start and end dates in that order, \emph{e.g.},
-\code{dates = c("1983-01-01", "2017-12-31")}.  See argument details for more.}
+\code{dates = c("1983-01-01", "2017-12-31")}.  Not required for global
+coverage.  See argument details for more.}
 
 \item{temporal_average}{Temporal average for data being queried, currently
 supported are DAILY, INTERANNUAL, CLIMATOLOGY.  See argument details for
