### 2018-06-03 ae405d2
check latlon for global as first step
This affects dates and temporal average downstream
