### 2018-06-03 fcfaa8f
Work on fixing tests
