+++ tests/testthat/test-get_power.R
@@ -2,7 +2,7 @@
 # test queries -----------------------------------------------------------------
 context("Test that get_power queries the server and returns the proper
         requested data")
-test_that("get_power returns AG data for a single point", {
+test_that("get_power returns daily point AG data", {
   vcr::use_cassette("SinglePoint_AG", {
     query <- get_power(
       community = "AG",
@@ -29,7 +29,7 @@ test_that("get_power returns AG data for a single point", {
   })
 })
 
-test_that("get_power returns AG data for a region", {
+test_that("get_power returns daily regional AG data", {
   vcr::use_cassette("Regional_AG", {
     query <- get_power(
       community = "AG",
@@ -37,6 +37,42 @@ test_that("get_power returns AG data for a region", {
       pars =  "T2M",
       dates = c("1983-01-01"),
       temporal_average = "Daily"
+    )
+
+    expect_equal(nrow(query), 77)
+    expect_equal(
+      unique(query$LAT),
+      c(
+        -55.25,
+        -54.75,
+        -54.25,
+        -53.75,
+        -53.25,
+        -52.75,
+        -52.25,
+        -51.75,
+        -51.25,
+        -50.75,
+        -50.25
+      )
+    )
+    expect_equal(unique(query$LON),
+                 c(112.75, 113.25, 113.75, 114.25, 114.75, 115.25, 115.75))
+    expect_equal(query$YEAR[1], 1983)
+    expect_equal(query$DOY[1], "001")
+    expect_equal(query$T2M[1], 3.28)
+    rm(query)
+  })
+})
+
+test_that("get_power returns global AG data for climatology", {
+  vcr::use_cassette("Global_AG", {
+    query <- get_power(
+      community = "AG",
+      latlon = "Global",
+      pars =  "T2M",
+      dates = c("1983-01-01"),
+      temporal_average = "Climatology"
     )
 
     expect_equal(nrow(query), 77)
