### 2018-06-04 48ef788
Add tests for Global area
