+++ tests/testthat/test-get_power.R
@@ -71,32 +71,36 @@ test_that("get_power returns global AG data for climatology", {
       community = "AG",
       latlon = "Global",
       pars =  "T2M",
-      dates = c("1983-01-01"),
       temporal_average = "Climatology"
     )
 
-    expect_equal(nrow(query), 77)
-    expect_equal(
-      unique(query$LAT),
+    expect_equal(nrow(query), 259200)
+    expect_equal(query$PARAMETER[1], "T2M")
+    expect_equal(query$LAT[259200], 89.75)
+    expect_equal(query$LAT[1], -89.75)
+    expect_equal(query$LON[259200], 179.75)
+    expect_equal(query$LON[1], -179.75)
+    expect_named(
+      query,
       c(
-        -55.25,
-        -54.75,
-        -54.25,
-        -53.75,
-        -53.25,
-        -52.75,
-        -52.25,
-        -51.75,
-        -51.25,
-        -50.75,
-        -50.25
+        "LAT",
+        "LON",
+        "PARAMETER",
+        "JAN",
+        "FEB",
+        "MAR",
+        "APR",
+        "MAY",
+        "JUN",
+        "JUL",
+        "AUG",
+        "SEP",
+        "OCT",
+        "NOV",
+        "DEC",
+        "ANN"
       )
     )
-    expect_equal(unique(query$LON),
-                 c(112.75, 113.25, 113.75, 114.25, 114.75, 115.25, 115.75))
-    expect_equal(query$YEAR[1], 1983)
-    expect_equal(query$DOY[1], "001")
-    expect_equal(query$T2M[1], 3.28)
     rm(query)
   })
 })
