### 2018-06-04 b725002
Update tests to use new functionality
