### 2018-06-04 00cd35a
Correct testthat item
no `check_latlon_cell` function in package any longer
