### 2018-06-04 8718b11
Add checks for Global requests and format messages
