+++ R/internal_functions.R
@@ -5,70 +5,89 @@
 }
 
 #' @noRd
-check_dates <- function(dates) {
-  # if dates are NULL, set to defaults
-  if (is.null(dates)) {
-    dates <- c("1983-01-01", as.character(Sys.Date()))
+# start/end dates not required for global.
+check_global <- function(latlon) {
+  if (is.character(latlon)) {
+    latlon <- paste0(toupper(substr(latlon, 1, 1)),
+                     tolower(substr(latlon, 2, nchar(latlon))))
+    if (latlon != "Global") {
+      stop("\nYou have entered an invalid value for `latlon`.\n")
+    }
   }
-  if (length(dates) == 1) {
-    dates <- c(dates, dates)
+  return(latlon)
+}
+
+#' @noRd
+check_dates <- function(dates, latlon) {
+  if (is.numeric(latlon)) {
+    # if dates are NULL, set to defaults
+    if (is.null(dates)) {
+      dates <- c("1983-01-01", as.character(Sys.Date()))
+    }
+    if (length(dates) == 1) {
+      dates <- c(dates, dates)
+    }
+
+    if (length(dates) > 2) {
+      stop(call. = FALSE,
+           "\nYou have supplied more than two dates for a start and end date.\n")
+    }
+
+    # put dates in list to use lapply
+    dates <- as.list(dates)
+
+    # check dates as entered by user
+    date_format <- function(x) {
+      tryCatch(
+        # try to parse the date format using lubridate
+        x <- lubridate::parse_date_time(x,
+                                        c(
+                                          "Ymd",
+                                          "dmY",
+                                          "mdY",
+                                          "BdY",
+                                          "Bdy",
+                                          "bdY",
+                                          "bdy"
+                                        )),
+        warning = function(c) {
+          stop(call. = FALSE,
+               "\n", x,
+               " is not a valid entry for date. Enter as YYYY-MM-DD.\n")
+        }
+      )
+      return(as.Date(x))
+    }
+
+    # apply function to reformat/check dates
+    dates = lapply(X = dates, FUN = date_format)
+
+    # if the stdate is > endate, flip order
+    if (dates[[2]] < dates[[1]]) {
+      message("\nYour start and end dates were reversed.\n",
+              "They have been reordered.\n")
+      dates <- as.list(c(dates[2], dates[1]))
+    }
+
+    # check date to be sure it's not before POWER data start
+    if (dates[[1]] < "1983-01-01") {
+      stop(call. = FALSE,
+           "\nNASA-POWER data do not start before 1983-01-01.\n")
+    }
+
+    # check end date to be sure it's not _after_
+    if (dates[[2]] > Sys.Date()) {
+      stop(call. = FALSE,
+           "\nThe data cannot possibly extend beyond this moment.\n")
+    }
+
+    dates <- lapply(dates, as.character)
+    dates <- gsub("-", "" , dates, ignore.case = TRUE)
+  } else if (is.character(latlon) & !is.null(dates)) {
+    message("\nDates are not used with CLIMATOLOGY. Setting to NULL.\n")
+    dates <- NULL
   }
 
-  if (length(dates) > 2) {
-    stop(call. = FALSE,
-         "You have supplied more than two dates for a start and end date.")
-  }
-
-  # put dates in list to use lapply
-  dates <- as.list(dates)
-
-  # check dates as entered by user
-  date_format <- function(x) {
-    tryCatch(
-      # try to parse the date format using lubridate
-      x <- lubridate::parse_date_time(x,
-                                      c(
-                                        "Ymd",
-                                        "dmY",
[…]
