### 2018-06-04 9e62453
Add a bit more detail to README
