+++ README.Rmd
@@ -26,7 +26,8 @@ available for download through a web interface at a resolution of 0.5˚ longitud
 by 0.5˚ latitude.
 
 Please see
-<https://power.larc.nasa.gov/> for more on the data and other ways to access it.
+<https://power.larc.nasa.gov/> for more on the data and other ways to access it
+and other forms of data available, e.g. an ESRI REST API.
 
 ### Quick start
 
