### 2018-06-04 5040729
Add citation message to startup
