### 2018-06-04 3eaa070
Remove tests for deprecated functions
