+++ tests/testthat/test-deprecated.R
@@ -1,9 +1,0 @@
-
-# test deprecated functions return a warning -----------------------------------
-context("Test that deprecated fucntions emit warning")
-test_that("deprecated functions return warning message", {
-  expect_warning(get_cell(),
-                 regexp = "'get_cell' is deprecated.*")
-  expect_warning(get_region(),
-                 regexp = "'get_region' is deprecated.*")
-})
