### 2018-06-04 7c77fd1
Update documentation
remove links to removed functions
