### 2018-06-04 f72e8cb
Remove Aji from authors' list
Did not contribute to this version of `nasapower`
