### 2018-06-05 6d4f18d
Add data-raw skeleton
