### 2018-06-12 ddc05b0
Update data-raw directory
Add directory to .Rbuildignore
