### 2018-06-22 1bbea71
Add params.rda list of parameters as data.file for checking
