+++ data-raw/README.Rmd
@@ -21,7 +21,7 @@ for pointing me to this file.
 Using `jsonlite` read the JSON file into R creating a list.
 
 ```{r fetch_JSON}
-params <-
+parameters <-
   jsonlite::fromJSON(
     "https://power.larc.nasa.gov/RADAPP/GEODATA/powerWeb/POWER_Parameters_v108.json"
   )
@@ -33,7 +33,7 @@ Using `devtools` to save the list as an R data object for use in the
 `nasapower` package.
 
 ```{r save_list, message=FALSE}
-devtools::use_data(params, overwrite = TRUE)
+devtools::use_data(parameters, overwrite = TRUE)
 ```
 
 ## Session Info
