### 2018-06-22 02701a3
Update documentation for parameters list
