+++ R/get_power.R
@@ -1,31 +1,31 @@
 
-#'@title Get NASA-POWER Data and Return a Tidy Data Frame
+#' @title Get NASA-POWER Data and Return a Tidy Data Frame
 #'
-#'@description Get NASA-POWER data and return a tidy data frame. All options
-#'offered by the official NASA-POWER API are supported.
+#' @description Get NASA-POWER data and return a tidy data frame. All options
+#' offered by the official NASA-POWER API are supported.
 #'
-#'@export
-#'@param community Dataset name, currently supported are AG, SB and SSE. See
+#' @export
+#' @param community Dataset name, currently supported are AG, SB and SSE. See
 #'  argument details for more.
-#'@param latlon A numeric vector of geographic coordinates for a cell or region
+#' @param latlon A numeric vector of geographic coordinates for a cell or region
 #'  entered as x, y coordinates or `Global` for global area.  See argument
 #'  details for more.
-#'@param pars A character vector of solar or meteorological variables to
+#' @param pars A character vector of solar or meteorological variables to
 #'  download.  See the `Value` field of [`parameters`] for a full list of valid
 #'  values or visit the [POWER website](https://power.larc.nasa.gov/new/) for
 #'  the Parameter Dictionary.  If downloading global coverage a maximum of 3
 #'  `pars` can be specified at one time.
-#'@param dates A character vector of start and end dates in that order, *e.g.*,
+#' @param dates A character vector of start and end dates in that order, *e.g.*,
 #'  `dates = c("1983-01-01", "2017-12-31")`.  Not required for global
 #'  coverage.  See argument details for more.
-#'@param temporal_average Temporal average for data being queried, currently
+#' @param temporal_average Temporal average for data being queried, currently
 #'  supported are DAILY, INTERANNUAL, CLIMATOLOGY.  See argument details for
 #'  more.
 #'
-#'@details Further details for each of the arguments are provided in their
+#' @details Further details for each of the arguments are provided in their
 #'respective sections following below.
 #'
-#'@section Argument details for `community`: There are three valid values, one
+#' @section Argument details for `community`: There are three valid values, one
 #'  must be supplied. This  will affect the units of the parameter and the
 #'  temporal display of time series data.
 #'
@@ -42,7 +42,7 @@
 #'  powered renewable energy systems.}
 #'  }
 #'
-#'@section Argument details for `latlon`:
+#' @section Argument details for `latlon`:
 #' \describe{
 #'  \item{For a single point}{To get a specific cell, 1/2 x 1/2 degree, supply a
 #'  length-2 numeric vector giving the decimal degree longitude and latitude in
@@ -61,35 +61,35 @@
 #'  `climatology` if this option is set.}
 #' }
 #'
-#'@section Argument details for `dates`: If `dates` is unspecified, defaults to
+#' @section Argument details for `dates`: If `dates` is unspecified, defaults to
 #'  a start date of 1983-01-01 (the earliest available data) and an end date of
 #'  current date according to the system.
 #'
 #'  If one date only is provided, it will be treated as both the start date and
 #'  the end date and only a single day's values will be returned.
 #'
-#'@section Argument details for `temporal_average`: There are three valid
+#' @section Argument details for `temporal_average`: There are three valid
 #'  values, one must be supplied. \describe{ \item{DAILY}{The daily average of
 #'  `pars` by year.} \item{INTERANNUAL}{The monthly average of `pars` by year.}
 #'  \item{CLIMATOLOGY}{The monthly average of `pars` at the surface of the earth
[…]
