+++ man/get_power.Rd
@@ -100,10 +100,10 @@ for a given month, averaged for that month over the 30-year period (Jan.
 
 \dontrun{
 power <- get_power(community = "AG",
-                  latlon = c(-179.5, -89.5),
-                  pars = c("RH2M", "T2M"),
-                  dates = "1985-01-01",
-                  temporal_average = "daily")
+                   latlon = c(-179.5, -89.5),
+                   pars = c("RH2M", "T2M"),
+                   dates = "1985-01-01",
+                   temporal_average = "daily")
 }
 
 }
@@ -111,5 +111,5 @@ power <- get_power(community = "AG",
 \url{https://power.larc.nasa.gov/documents/Agroclimatology_Methodology.pdf}
 }
 \author{
-Adam H. Sparks, adamhsparks@gmail.com
+Adam H. Sparks, \email{adamhsparks@gmail.com}
 }
