+++ man/nasapower.Rd
@@ -17,7 +17,6 @@ themselves, please see \url{https://power.larc.nasa.gov/}.
 \note{
 While \pkg{nasapower} does not redistribute the data in any way, we
 encourage users to follow the requests of the POWER Project Team.
-
 \preformatted{
 When POWER data products are used in a publication, we request the
 following acknowledgment be included:
