### 2018-06-22 8a1adf5
Remove duplicate author field in documentation for package
