### 2018-06-22 5755c00
Replace NASA/POWER with NASA-POWER
