### 2018-06-22 4f4eb86
Update internal functions to use parameters list from JSON file
