### 2018-06-22 4360d23
Fix typo in internal functions
