### 2018-06-22 e19de77
Rename hex to logo.png and rebuild site
