+++ README.Rmd
@@ -17,7 +17,7 @@ library(nasapower)
 
 ## Introduction
 
-<img align="right" src="man/figures/nasapower-hex.png"> _nasapower_ aims to make
+<img align="right" src="man/figures/logo.png"> _nasapower_ aims to make
 it quick and easy to automate downloading
 [NASA-POWER](https://power.larc.nasa.gov) agroclimatology data in your R session
 as a tidy data frame for agricultural analysis and use in modelling or other
