### 2018-06-22 #155 @adamhsparks
OK! After several months hiatus, I think we're back in business.

I've **completely** rewritten the `nasapower` package to utilise the new NASA-POWER API that was recently released and use the new POWER data. In essence, this is a new R package. If there is any code carryover it's extremely minimal due to the new processes and interfaces. I did manage to keep one vignette, the use-case, the same though.

Unfortunately due to the way the API handles requests, I can't fetch a whole country like I could with the previous version (in the vignette). This may be possible if I (or someone else smarter than me) can figure out how to access the ESRI REST API. For now you can query a single point, a region in a 5x5 arc degree area for daily data or global climatology data.

I think I'm ready for the editors to decide what to do with this now that it's been reviewed once and completely rewritten due to changes in the source immediately after the original reviews.
