### 2018-06-22 fe83d26
Address goodpractice suggestions where appropriate
