+++ tests/testthat/test-check_community.R
@@ -1,6 +1,7 @@
 
 # community checks -------------------------------------------------------------
-context("Test that check_community function handles community strings correctly")
+context("Test that check_community function handles community strings
+        correctly")
 test_that("check_community fails if one not provided", {
   expect_error(check_community(NULL),
                regexp = "*You have not provided a `community` value.*")
