### 2018-06-22 81ecd4d
Clean up NEWS.md
