### 2018-06-22 e6ea47a
Correct tests so they pass
