### 2018-06-22 b6c5bb6
httr is not used, remove from Imports
