### 2018-06-22 9b8c00d
Shorten line lengths in tests
