### 2018-06-23 89a82f8
Fix more tests
