+++ tests/testthat/test-check_pars.R
@@ -30,7 +30,7 @@ test_that("check_pars stops if `temporal_average` not valid", {
       temporal_average = "ad09jlave",
       latlon = c(-89.5, -179.5)
     ),
-    regexp = "\nYou have supplied more than two dates for start and end*")
+    regexp = "\nYou have entered an invalid value for `temporal_average`.\n")
 })
 
 test_that("check_pars stops if `pars` not valid", {
@@ -50,8 +50,7 @@ test_that("check_pars stops if `pars` not valid", {
       temporal_average = "Interannual",
       latlon = c(-89.5, -179.5)
     ),
-    regexp = "*You have entered an invalid value for `temporal_average` for*"
-  )
+    regexp = "*You have entered an invalid value for `temporal_average` for*")
 })
 
 test_that("pars are returned as a comma separated string with no spaces", {
