+++ DESCRIPTION
@@ -7,13 +7,14 @@ Authors@R: person(given = "Adam H.", family = "Sparks",
     comment = c(ORCID = "0000-0002-0061-8359"))
 URL: https://github.com/adamhsparks/nasapower
 BugReports: https://github.com/adamhsparks/nasapower/issues
-Description: Download NASA-POWER agroclimatology data and create a tidy data
-    frame for individual cells or a user specified region. nasapower provides an
-    R interface to the NASA - POWER API.  POWER (Prediction Of Worldwide Energy
-    Resource) data are freely available for download with a resolution of 1/2 by
-    1/2 arc degree longitude and latitude and are funded through the NASA
-    Earth Science Directorate Applied Science Program. For more on the data
-    themselves, please see <https://power.larc.nasa.gov/>.
+Description: Download NASA-POWER meteorology and agroclimatology data and create
+    a tidy data frame for individual cells or a user specified region for 140
+    parameters.  nasapower provides an R interface to the NASA - POWER API.
+    POWER (Prediction Of Worldwide Energy Resource) data are freely available
+    for download with a resolution of 1/2 by 1/2 arc degree longitude and
+    latitude and are funded through the NASA Earth Science Directorate Applied
+    Science Program. For more on the data themselves, please see
+    <https://power.larc.nasa.gov/>.
 Depends:
     R (>= 3.2.0)
 License: MIT + file LICENSE
