### 2018-06-23 b483120
Replace degree symbol and fix number of params to 141
