+++ R/internal_functions.R
@@ -1,5 +1,4 @@
 
-
 `%notin%` <- function(x, table) {
   # Same as !(x %in% table)
   match(x, table, nomatch = 0L) == 0L
