+++ README.Rmd
@@ -22,8 +22,8 @@ it quick and easy to automate downloading
 [NASA-POWER](https://power.larc.nasa.gov) agroclimatology data in your R session
 as a tidy data frame for agricultural analysis and use in modelling or other
 purposes. POWER (Prediction Of Worldwide Energy Resource) data are freely
-available for download through a web interface at a resolution of 0.5˚ longitude
-by 0.5˚ latitude.
+available for download through a web interface at a resolution of 0.5 arc degree
+longitude by 0.5 arc degree latitude.
 
 Please see
 <https://power.larc.nasa.gov/> for more on the data and other ways to access it
