### 2018-06-24 cb40856
Replace UTF-8 chars in included data
