### 2018-06-24 c7f9d2c
Correct replacement of UTF8 chars where I goofed
