+++ data-raw/README.Rmd
@@ -30,10 +30,10 @@ parameters <-
 Replace UTF-8 characters in the dataset since R doesn't like this in packages.
 
 ```{r, replace_utf}
-parameters$ALLSKY_SFC_LW_DWN$climatology_definition <-
+parameters$SG_DEC_AVG$climatology_definition <-
   gsub("°",
        " degrees",
-       parameters$SG_HR_SET_ANG$climatology_definition)
+       parameters$SG_DEC_AVG$climatology_definition)
 
 parameters$SG_HR_SET_ANG$climatology_definition <-
   gsub("°",
