### 2018-06-24 b2353b3
Update links and documentation for parameters
