+++ R/get_power.R
@@ -12,9 +12,10 @@
 #'  details for more.
 #' @param pars A character vector of solar, meteorological or climatology
 #'  parameters to download.  See \code{names(parameters)} for a full list of
-#'  valid values or visit the [POWER website](https://power.larc.nasa.gov/) for
-#'  the Parameter Dictionary.  If downloading global coverage a maximum of 3
-#'  `pars` can be specified at one time.
+#'  valid values and definitions visit the
+#'  [POWER website](https://power.larc.nasa.gov/#resources) for the Parameter
+#'  Dictionary.  If downloading global coverage a maximum of 3 `pars` can be
+#'  specified at one time.
 #' @param dates A character vector of start and end dates in that order,\cr
 #'  *e.g.*, `dates = c("1983-01-01", "2017-12-31")`.  Not required for global
 #'  coverage.  See argument details for more.
