+++ vignettes/nasapower.Rmd
@@ -55,11 +55,10 @@ a region, the maximum amount of cells that can be downloaded are 100, i.e., a 10
 x 10 one-half arc degree bounding box or 5 x 5 one arc degree bounding box.
 Leave blank for global coverage.
 
-- `pars`, A character vector of solar or meteorological variables to download.
-See the `Value` field of [`parameters`](../help/parameters) for a full list of
-valid values or visit the [POWER website](https://power.larc.nasa.gov/new/) for
-the Parameter Dictionary. If downloading global coverage a maximum of 3 `pars`
-can be specified at one time.
+- `pars`, A character vector of solar, meteorological or climatology parameters
+to download. For a complete listing of valid `pars`, please visit the NASA-POWER
+website for the dictionary of parameters,
+<https://power.larc.nasa.gov/#resources>.
 
 - `dates`, a vector of start and end dates for which to query the NASA-POWER
 API, *e.g.*, `dates = c("1983-01-01", "2017-12-31")`. 
