### 2018-06-25 e22c27d
Delete an extra line
