### 2018-06-25 a43bf92
Update internal functions message
