+++ R/internal_functions.R
@@ -361,7 +361,7 @@ power_query <- function(community,
     e$message <-
       paste("\nA CSV file was not created, this is a server error.",
             "The server may not be responding.",
-            "Please check https://power.larc.nasa.gov/ for notifications if",
+            "Please check <https://power.larc.nasa.gov/> for notifications if",
             "you repeatedly get this error.", sep = "\n")
     # Otherwise refers to open.connection
     e$call <- NULL
