### 2018-06-25 #155 @sckott
Thanks @adamhsparks for the update.

I dropped the status back to 3, reviewers assigned, I think we should get new reviews. Looks like @UniversalTourist already said she can review again. @alisonboyer could you review this new version of the package?  If not just let me know
