### 2018-06-25 #155 @adamhsparks
Very good, I agree, it needs new reviews since it's basically a new package.

Thanks for your patience, @sckott.
