### 2018-06-27 627775a
Fix formatting of some of the documentation for consistency
