### 2018-06-27 6243008
Correct code for regional latlon example
Remove extra ","
