### 2018-06-27 db17604
Add new function to add dates to final data.frame `format_dates`
This adds the fields for YYYY (int), MM (int), DD (int), YYYYMMDD (date)
