### 2018-06-27 ef2f5d1
Fix bugs in format_dates
