### 2018-06-27 30f9ef2
Check if global average is CLIMATOLOGY before formatting dates
