### 2018-06-27 1df3d76
Add comments to get_power.R
