### 2018-06-27 c34f3f5
Add function to check agreement between temporal average and dates
Needs further work
