### 2018-06-27 03f66b9
Update documentation in vignette to match current functionality
