### 2018-06-27 4c7f2ff
(Incomplete) Update paper to match current functionality
