### 2018-06-27 4675c3b
Incorporate temporal average and date checks into check_dates
