### 2018-06-27 bbc4122
Fix check for CLIMATOLOGY and error message
