### 2018-06-27 03122c1
Add test to check for temporal average in check_dates()
