### 2018-06-27 7b3c478
Update vignette example converting to raster
