### 2018-06-27 f8bb7e8
Fix bugs with INTERANNUAL checks and update message formats
