### 2018-06-27 cb85278
Add tests for INTERANNUAL checks
