+++ DESCRIPTION
@@ -22,7 +22,8 @@ Imports:
     crul,
     lubridate,
     jsonlite,
-    readr
+    readr,
+    tibble
 RoxygenNote: 6.0.1
 Roxygen: list(markdown = TRUE)
 Encoding: UTF-8
