### 2018-06-27 e15f7a3
Fix tests for check_dates()
