+++ tests/testthat/test-check_dates.R
@@ -1,5 +1,6 @@
 
 
+
 # Date handling and checking ---------------------------------------------------
 context("Check dates function handles dates correctly")
 test_that("NULL `dates`` are properly handled", {
@@ -26,7 +27,7 @@ test_that("`dates` >2 cause an error", {
   temporal_average <- "DAILY"
   dates <- c("1983-01-01", "1983-01-02", "1983-01-03")
   latlon <- c(-89.5, -179.5)
-  expect_error(check_dates(dates, latlon),
+  expect_error(check_dates(dates, latlon, temporal_average),
                regexp =
                  "\nYou have supplied more than two dates for start and end*")
 })
@@ -57,7 +58,7 @@ test_that("`dates` before the start of POWER data cause error", {
   today <- as.character(Sys.Date())
   dates <- c("1982-12-31", today)
   latlon <- c(-89.5, -179.5)
-  expect_error(check_dates(dates, latlon),
+  expect_error(check_dates(dates, latlon, temporal_average),
                regexp = "*NASA-POWER data do not start before 1983-01-01*")
 })
 
@@ -73,7 +74,7 @@ test_that("Invalid `dates` are handled", {
   temporal_average <- "DAILY"
   dates <- c("1983-01-01", "1983-02-31")
   latlon <- c(-89.5, -179.5)
-  expect_error(check_dates(dates, latlon),
+  expect_error(check_dates(dates, latlon, temporal_average),
                regexp = "*1983-02-31 is not a valid entry for date.*")
 })
 
@@ -85,15 +86,6 @@ test_that("Dates are returned as a vector of characters", {
   expect_is(dates, "character")
 })
 
-test_that("A `global` `latlon` returns `NULL` dates", {
-  temporal_average <- "DAILY"
-  latlon <- "Global"
-  dates <- c("1983-01-01", "1983-02-02")
-  expect_message(dates <- check_dates(dates, latlon, temporal_average),
-                 regexp = "*are not used with CLIMATOLOGY. Setting to NULL.\n")
-  expect_true(is.null(dates))
-})
-
 test_that("If temporal_average == CLIMATOLOGY, no dates can be specified", {
   temporal_average <- "CLIMATOLOGY"
   dates <- c("1983-01-01", "1983-02-02")
@@ -102,20 +94,26 @@ test_that("If temporal_average == CLIMATOLOGY, no dates can be specified", {
                regexp = "*Dates are not used when querying climatology data.*")
 })
 
-test_that("If temporal_average == INTERANNUAL and dates are specified, that
-          a message is emitted about years only", {
-  temporal_average <- "INTERANNUAL"
-  dates <- c("1983-01-01", "1984-01-01")
-  latlon <- c(-89.5, -179.5)
-  expect_message(check_dates(dates, latlon, temporal_average),
-               regexp = "*Only years are used with INTERANNUAL temporal.*")
-})
+test_that(
+  "If temporal_average == INTERANNUAL and dates are specified, that
+  a message is emitted about years only",
+  {
+    temporal_average <- "INTERANNUAL"
+    dates <- c("1983-01-01", "1984-01-01")
+    latlon <- c(-89.5, -179.5)
+    expect_message(check_dates(dates, latlon, temporal_average),
+                   regexp = "*Only years are used with INTERANNUAL temporal.*")
+  }
+)
 
-test_that("If temporal_average == INTERANNUAL and dates are specified, that
-          a message is emitted about years only", {
-            temporal_average <- "INTERANNUAL"
-            dates <- c("1983-01-01", "1983-02-02")
-            latlon <- c(-89.5, -179.5)
-            expect_error(check_dates(dates, latlon, temporal_average),
[…]
