### 2018-06-27 e27f59d
Update tests for get_power
