+++ tests/testthat/test-get_power.R
@@ -4,7 +4,7 @@ context("Test that get_power queries the server and returns the proper
         requested data")
 test_that("get_power returns daily point AG data", {
   vcr::use_cassette("SinglePoint_AG", {
-    query <- get_power(
+    power_query <- get_power(
       community = "AG",
       latlon = c(-89.5, -179.5),
       pars =  c("T2M",
@@ -19,19 +19,19 @@ test_that("get_power returns daily point AG data", {
     expect_equal(query$LAT, -89.5)
     expect_equal(query$LON, -179.5)
     expect_equal(query$YEAR, 1983)
-    expect_equal(query$DOY, "001")
+    expect_equal(query$DOY, 1)
     expect_equal(query$T2M, -25.24)
     expect_equal(query$T2M_MIN, -25.55)
     expect_equal(query$T2M_MAX, -24.9)
     expect_equal(query$RH2M, 73.92)
     expect_equal(query$WS10M, 2.14)
-    rm(query)
+    rm(power_query)
   })
 })
 
 test_that("get_power returns daily regional AG data", {
   vcr::use_cassette("Regional_AG", {
-    query <- get_power(
+    power_query <- get_power(
       community = "AG",
       latlon = c(-55.5, 112.5, -50.5, 115.5),
       pars =  "T2M",
@@ -59,7 +59,7 @@ test_that("get_power returns daily regional AG data", {
     expect_equal(unique(query$LON),
                  c(112.75, 113.25, 113.75, 114.25, 114.75, 115.25, 115.75))
     expect_equal(query$YEAR[1], 1983)
-    expect_equal(query$DOY[1], "001")
+    expect_equal(query$DOY[1], 1)
     expect_equal(query$T2M[1], 3.28)
     rm(query)
   })
@@ -67,7 +67,7 @@ test_that("get_power returns daily regional AG data", {
 
 test_that("get_power returns global AG data for climatology", {
   vcr::use_cassette("Global_AG", {
-    query <- get_power(
+    power_query <- get_power(
       community = "AG",
       latlon = "Global",
       pars =  "T2M",
