### 2018-06-27 abb4818
Clean up Suggests
Remove libraries that are not used
