+++ DESCRIPTION
@@ -39,11 +39,8 @@ Suggests:
     knitr,
     raster,
     rmarkdown,
-    rnaturalearth,
-    sp,
     testthat,
-    vcr,
-    viridis
+    vcr
 VignetteBuilder: knitr
 X-schema.org-applicationCategory: Tools
 X-schema.org-keywords: NASA, meteorological-data, weather, global, weather, weather-data, meteorology, NASA-POWER, agroclimatology, earth-science, data-access
