### 2018-06-27 0919174
Put lon before lat in data frame
