### 2018-06-27 84498ab
Add figures to nasapower vignette and set vignettes to not execute any code for faster package builds
