### 2018-06-28 91d8735
Clean up headings in vignette
