### 2018-06-28 18afc41
Correct test for climatology, I reordered fields lon preceeds lat now
