+++ inst/paper/paper.bib
@@ -1,3 +1,4 @@
+% Encoding: UTF-8
 
 @manual{fainges2017,
 	Author = {Justin Fainges},
@@ -9,16 +10,16 @@
 	Url = {https://CRAN.R-project.org/package=APSIM},
 	Year = {2017}}
 
-@webpage{stackhouse2017,
-	Author = {Paul W. {Stackhouse, Jr} and David Westberg and William S. Chandler and and Taiping Zhang and James M. Hoell},
-	Date-Added = {2017-12-21 05:54:40 +0000},
-	Date-Modified = {2017-12-21 05:58:20 +0000},
-	Lastchecked = {2017-12-21},
-	Month = {05},
-	Title = {Prediction Of Worldwide Energy Resource (POWER) Agroclimatology Methodology (1.0 Latitude by 1.0 Longitude Spatial Resolution)},
-	Url = {https://power.larc.nasa.gov/documents/Agroclimatology_Methodology.pdf},
-	Urldate = {2017-05-30},
-	Year = {2017}}
+@TechReport{stackhouseJr2018,
+  author      = {Paul W. {Stackhouse, Jr} and Taiping Zhang and David Westberg and A. Jason Barnett and Tyler Bristow and Bradley Macpherson and James M. Hoell},
+  title       = {POWER Release 8 (with {GIS} Applications) Methodology (Data Parameters, Sources, \& Validation) Documentation Date May 1, 2018 (All previous versions are obsolete) (Data Version 8.0.1)},
+  institution = {NASA},
+  year        = {2018},
+  date        = {2018-05},
+  lastchecked = {2018-06-28},
+  url         = {https://power.larc.nasa.gov/documents/POWER_Data_v8_methodology.pdf},
+  urldate     = {2018-05-01},
+}
 
 @article{bai2010evaluation,
 	Author = {Bai, Jinshun and Chen, Xinping and Dobermann, Achim and Yang, Haishun and Cassman, Kenneth G and Zhang, Fusuo},
@@ -53,6 +54,14 @@
 	Year = {2003},
 	Bdsk-Url-1 = {http://www.sciencedirect.com/science/article/pii/S1161030102001089},
 	Bdsk-Url-2 = {https://doi.org/10.1016/S1161-0301(02)00108-9}}
+
+  @Manual{muller2018,
+    title = {tibble: Simple Data Frames},
+    author = {Kirill Müller and Hadley Wickham},
+    year = {2018},
+    note = {R package version 1.4.2},
+    url = {https://CRAN.R-project.org/package=tibble},
+  }
 
 @article{ojeda2017,
 	Author = {Ojeda, Jonathan J and Volenec, Jeffrey J and Brouder, Sylvie M and Caviglia, Octavio P and Agnusdei, M{\'o}nica G},
@@ -149,3 +158,5 @@
 	Year = {2014},
 	Bdsk-Url-1 = {https://www.jstatsoft.org/index.php/jss/article/view/v059i10},
 	Bdsk-Url-2 = {https://dx.doi.org/10.18637/jss.v059.i10}}
+
+@Comment{jabref-meta: databaseType:bibtex;}
