### 2018-06-28 ee260ff
Add DSSAT citations
