+++ inst/paper/paper.bib
@@ -1,14 +1,15 @@
 % Encoding: UTF-8
 
-@manual{fainges2017,
-	Author = {Justin Fainges},
-	Date-Added = {2017-12-21 06:17:25 +0000},
-	Date-Modified = {2017-12-21 06:20:15 +0000},
-	Note = {R package version 0.9.2},
-	Organization = {CSIRO},
-	Title = {{APSIM}: General Utility Functions for the 'Agricultural Production Systems Simulator'},
-	Url = {https://CRAN.R-project.org/package=APSIM},
-	Year = {2017}}
+@Manual{fainges2017,
+  title         = {{APSIM}: General Utility Functions for the 'Agricultural Production Systems Simulator'},
+  author        = {Justin Fainges},
+  organization  = {CSIRO},
+  note          = {R package version 0.9.2},
+  date          = {2017},
+  date-added    = {2017-12-21 06:17:25 +0000},
+  date-modified = {2017-12-21 06:20:15 +0000},
+  url           = {https://CRAN.R-project.org/package=APSIM},
+}
 
 @TechReport{stackhouseJr2018,
   author      = {Paul W. {Stackhouse, Jr} and Taiping Zhang and David Westberg and A. Jason Barnett and Tyler Bristow and Bradley Macpherson and James M. Hoell},
@@ -21,142 +22,172 @@
   urldate     = {2018-05-01},
 }
 
-@article{bai2010evaluation,
-	Author = {Bai, Jinshun and Chen, Xinping and Dobermann, Achim and Yang, Haishun and Cassman, Kenneth G and Zhang, Fusuo},
-	Journal = {Agronomy Journal},
-	Number = {1},
-	Pages = {9--16},
-	Publisher = {American Society of Agronomy},
-	Title = {Evaluation of NASA satellite-and model-derived weather data for simulation of maize yield potential in China},
-	Volume = {102},
-	Year = {2010}}
+@Article{bai2010evaluation,
+  author       = {Bai, Jinshun and Chen, Xinping and Dobermann, Achim and Yang, Haishun and Cassman, Kenneth G and Zhang, Fusuo},
+  title        = {Evaluation of NASA satellite-and model-derived weather data for simulation of maize yield potential in China},
+  volume       = {102},
+  number       = {1},
+  pages        = {9--16},
+  date         = {2010},
+  journaltitle = {Agronomy Journal},
+  publisher    = {American Society of Agronomy},
+}
 
-@manual{hijmans2017,
-	Author = {Robert J. Hijmans},
-	Note = {R package version 2.6-7},
-	Title = {{raster}: Geographic Data Analysis and Modeling},
-	Url = {https://CRAN.R-project.org/package=raster},
-	Year = {2017},
-	Bdsk-Url-1 = {https://CRAN.R-project.org/package=raster}}
+@Manual{hijmans2017,
+  title      = {{raster}: Geographic Data Analysis and Modeling},
+  author     = {Robert J. Hijmans},
+  note       = {R package version 2.6-7},
+  bdsk-url-1 = {https://CRAN.R-project.org/package=raster},
+  date       = {2017},
+  url        = {https://CRAN.R-project.org/package=raster},
+}
 
-@article{keating2003,
-	Author = {B.A Keating and P.S Carberry and G.L Hammer and M.E Probert and M.J Robertson and D Holzworth and N.I Huth and J.N.G Hargreaves and H Meinke and Z Hochman and G McLean and K Verburg and V Snow and J.P Dimes and M Silburn and E Wang and S Brown and K.L Bristow and S Asseng and S Chapman and R.L McCown and D.M Freebairn and C.J Smith},
-	Doi = {https://doi.org/10.1016/S1161-0301(02)00108-9},
[…]
