### 2018-06-28 2421498
Clean up and update bibliography
