+++ inst/paper/paper.bib
@@ -24,7 +24,7 @@
 
 @Article{bai2010evaluation,
   author       = {Bai, Jinshun and Chen, Xinping and Dobermann, Achim and Yang, Haishun and Cassman, Kenneth G and Zhang, Fusuo},
-  title        = {Evaluation of NASA satellite-and model-derived weather data for simulation of maize yield potential in China},
+  title        = {Evaluation of {NASA} satellite-and model-derived weather data for simulation of maize yield potential in China},
   volume       = {102},
   number       = {1},
   pages        = {9--16},
@@ -43,8 +43,8 @@
 }
 
 @Article{keating2003,
-  author       = {B.A Keating and P.S Carberry and G.L Hammer and M.E Probert and M.J Robertson and D Holzworth and N.I Huth and J.N.G Hargreaves and H Meinke and Z Hochman and G McLean and K Verburg and V Snow and J.P Dimes and M Silburn and E Wang and S Brown and K.L Bristow and S Asseng and S Chapman and R.L McCown and D.M Freebairn and C.J Smith},
-  title        = {An overview of APSIM, a model designed for farming systems simulation},
+  author       = {BA Keating and PS Carberry and GL Hammer and ME Probert and MJ Robertson and D Holzworth and NI Huth and JNG Hargreaves and H Meinke and Z Hochman and G McLean and K Verburg and V Snow and JP Dimes and M Silburn and E Wang and S Brown and KL Bristow and S Asseng and S Chapman and RL McCown and DM Freebairn and CJ Smith},
+  title        = {An overview of {APSIM}, a model designed for farming systems simulation},
   volume       = {18},
   number       = {3},
   pages        = {267 - 288},
@@ -60,7 +60,7 @@
 }
 
 @Manual{muller2018,
-  title  = {tibble: Simple Data Frames},
+  title  = {{tibble}: Simple Data Frames},
   author = {Kirill Müller and Hadley Wickham},
   note   = {R package version 1.4.2},
   date   = {2018},
@@ -69,7 +69,7 @@
 
 @Article{ojeda2017,
   author       = {Ojeda, Jonathan J and Volenec, Jeffrey J and Brouder, Sylvie M and Caviglia, Octavio P and Agnusdei, M{\'o}nica G},
-  title        = {Evaluation of Agricultural Production Systems Simulator as yield predictor of Panicum virgatum and Miscanthus x giganteus in several US environments},
+  title        = {{Evaluation of Agricultural Production Systems Simulator as Yield Predictor of \textit{Panicum virgatum} and \textit{Miscanthus} x \textit{giganteus} in Several {US} Environments}},
   volume       = {9},
   number       = {4},
   pages        = {796--816},
@@ -104,12 +104,12 @@
   url          = {http://www.sciencedirect.com/science/article/pii/S0261219411003656},
 }
 
-@Manual{sparks2017,
+@Manual{sparks2018,
   title      = {{nasapower}: NASA-POWER Agroclimatology Data from R},
   author     = {Adam H. Sparks},
-  note       = {R package version 0.1.3},
+  note       = {R package version 1.0.0},
   bdsk-url-1 = {https://CRAN.R-project.org/package=nasapower},
-  date       = {2017},
+  date       = {2018},
   url        = {https://CRAN.R-project.org/package=nasapower},
 }
 
@@ -172,7 +172,7 @@
 
 @Article{jones2003dssat,
   author       = {Jones, James W and Hoogenboom, Gerrit and Porter, Cheryl H and Boote, Ken J and Batchelor, William D and Hunt, LA and Wilkens, Paul W and Singh, Upendra and Gijsman, Arjan J and Ritchie, Joe T},
-  title        = {The DSSAT cropping system model},
[…]
