### 2018-06-28 cfeec8b
Spell check paper, update dates and citations
