+++ inst/paper/paper.md
@@ -10,6 +10,7 @@ tags:
 - weather data
 - meteorology
 - agroclimatology
+- climatology
 - R
 - climate data
 - earth science
@@ -17,46 +18,41 @@ tags:
 affiliations:
   index: 1
   name: University of Southern Queensland, Centre for Crop Health, Toowoomba Queensland 4350, Australia
-date: 2017-12-21
+date: 2018-28-06
 ---
 
 # Summary and Statement of Need
 
-_nasapower_ [@sparks2017] is an R [@R-base] package providing functionality to
-interface with the NASA-POWER API [@stackhouse2017] for reproducible data
-retrieval using R. A single function, `get_power` is provided. The function
+_nasapower_ [@sparks2018] is an R [@R-base] package providing functionality to
+interface with the NASA-POWER API [@stackhouseJr2018] for reproducible data
+retrieval using R. A single function, `get_power`, is provided. The function
 provides complete access to all functionality that the POWER API provides, which
-incluces three user communities, AG (agroclimatoloy), SSE (Surface meteorology
+includes three user communities, AG (agroclimatoloy), SSE (Surface meteorology
 and Solar Energy) and SB (Sustainable Buildings); three temporal averages, Daily
-Interannual and Climatology; three geographic options, single point, regional
+Inter-annual and Climatology; three geographic options, single point, regional
 and global for the appropriate 141 parameters offered. Data are returned in a
-in a tidy data frame [@wickham2014]. Integrating this data retrieval and
-formatting in R will streamline processes with models such as APSIM
-[@keating2003] and EPIRICE [@savary2012] that can be linked to or are
-implemented in R. Extended documentation is provided with examples of
-converting it to spatial objects using _raster_ [@hijmans2017] and for obtaining solar radiation values
-and generating a .met file using the _APSIM_ R package [@fainges2017] to use in
-the APSIM model.
+tidy data frame [@wickham2014] as a _tibble_ [@muller2018]. Integrating this
+data retrieval and formatting in R will streamline processes with models such as
+APSIM [@keating2003], DSSAT [@jones1998decision; @jones2003dssat] and EPIRICE
+[@savary2012] that can be linked to or are implemented fully in R. Extended
+documentation is provided with examples of converting it to spatial objects
+ using _raster_ [@hijmans2017] and for obtaining solar radiation values and
+ generating a .met file using the _APSIM_ R package [@fainges2017] to use in the
+ APSIM model.
 
 # About POWER Data
 
-NASA’s POWER (Prediction Of Worldwide Energy Resource) agroclimatology data
-[@stackhouse2017] are freely available for download via a
-[web interface](https://power.larc.nasa.gov/cgi-bin/agro.cgi?email=agroclim@larc.nasa.gov)
-at a grid resolution of one arc degree longitude by one arc degree latitude or
-roughly 111 square kilometres. Funded through the NASA Earth Science Directorate
-Applied Science Program, the data provide daily global coverage from 1983 until
-near present for all parameters except precipitation, which is provided for
-January 1997 to near present with a several month delay. Parameters available
-for download include: i) top-of-atmosphere insolation, ii) insolation on
-horizontal surface, iii) downward longwave radiative flux, iv) average air
-temperature at two meters, v) minimum air temperature at two meters, vi) maximum
-air temperature at two meters, vii) relative humidity at two meters, viii) dew
-point at two meters, ix) precipitation (January 1997 to near present with
-several month delay) and x) wind speed at ten meters. The data are widely used
-in agricultural modelling for modelling crop yields [@bai2010evaluation;
-@vanwart2013; @vanwart2015], other crop simulation exercises [@ojeda2017] and
-plant disease modelling [@savary2012].
[…]
