### 2018-06-28 d1a0fb2
Clean up spaces in paper.md
