### 2018-06-28 46eddfc
Clean up references in bib file
