+++ inst/paper/paper.bib
@@ -12,7 +12,7 @@
 }
 
 @TechReport{stackhouseJr2018,
-  author      = {Paul W. {Stackhouse, Jr} and Taiping Zhang and David Westberg and A. Jason Barnett and Tyler Bristow and Bradley Macpherson and James M. Hoell},
+  author      = {Paul W. {Stackhouse, Jr} and Taiping Zhang and David Westberg and A Jason Barnett and Tyler Bristow and Bradley Macpherson and James M Hoell},
   title       = {POWER Release 8 (with {GIS} Applications) Methodology (Data Parameters, Sources, \& Validation) Documentation Date May 1, 2018 (All previous versions are obsolete) (Data Version 8.0.1)},
   institution = {NASA},
   year        = {2018},
@@ -35,28 +35,22 @@
 
 @Manual{hijmans2017,
   title      = {{raster}: Geographic Data Analysis and Modeling},
-  author     = {Robert J. Hijmans},
+  author     = {Robert J Hijmans},
   note       = {R package version 2.6-7},
   bdsk-url-1 = {https://CRAN.R-project.org/package=raster},
   date       = {2017},
   url        = {https://CRAN.R-project.org/package=raster},
 }
 
-@Article{keating2003,
-  author       = {BA Keating and PS Carberry and GL Hammer and ME Probert and MJ Robertson and D Holzworth and NI Huth and JNG Hargreaves and H Meinke and Z Hochman and G McLean and K Verburg and V Snow and JP Dimes and M Silburn and E Wang and S Brown and KL Bristow and S Asseng and S Chapman and RL McCown and DM Freebairn and CJ Smith},
+@Article{keating2003overview,
+  author       = {Keating, Brian A and Carberry, Peter S and Hammer, Graeme L and Probert, Mervyn E and Robertson, Michael J and Holzworth, D and Huth, Neil I and Hargreaves, John NG and Meinke, Holger and Hochman, Zvi and others},
   title        = {An overview of {APSIM}, a model designed for farming systems simulation},
   volume       = {18},
-  number       = {3},
-  pages        = {267 - 288},
-  issn         = {1161-0301},
-  note         = {Modelling Cropping Systems: Science, Software and Applications},
-  bdsk-url-1   = {http://www.sciencedirect.com/science/article/pii/S1161030102001089},
-  bdsk-url-2   = {https://doi.org/10.1016/S1161-0301(02)00108-9},
+  number       = {3-4},
+  pages        = {267--288},
   date         = {2003},
-  doi          = {10.1016/S1161-0301(02)00108-9},
   journaltitle = {European Journal of Agronomy},
-  keywords     = {Farming systems modelling, APSIM, Simulation model},
-  url          = {http://www.sciencedirect.com/science/article/pii/S1161030102001089},
+  publisher    = {Elsevier},
 }
 
 @Manual{muller2018,
@@ -128,7 +122,7 @@
 }
 
 @Article{vanwart2015,
-  author       = {Justin {van Wart} and Patricio Grassini and Haishun Yang and Lieven Claessens and Andrew Jarvis and Kenneth G. Cassman},
+  author       = {Justin {van Wart} and Patricio Grassini and Haishun Yang and Lieven Claessens and Andrew Jarvis and Kenneth G Cassman},
   title        = {Creating long-term weather data from thin air for crop simulation modeling},
   volume       = {209-210},
   number       = {Supplement C},
