### 2018-06-28 472c84c
Clean up paper and citation
- Fix title

- Remove self-reference for package

- Add tags

- Split paper into two paragraphs
