+++ inst/paper/paper.bib
@@ -98,15 +98,6 @@
   url          = {http://www.sciencedirect.com/science/article/pii/S0261219411003656},
 }
 
-@Manual{sparks2018,
-  title      = {{nasapower}: NASA-POWER Agroclimatology Data from R},
-  author     = {Adam H. Sparks},
-  note       = {R package version 1.0.0},
-  bdsk-url-1 = {https://CRAN.R-project.org/package=nasapower},
-  date       = {2018},
-  url        = {https://CRAN.R-project.org/package=nasapower},
-}
-
 @Article{vanwart2013,
   author       = {{van Wart}, Justin and Grassini, Patricio and Cassman, Kenneth G.},
   title        = {Impact of derived global weather data on simulated crop yields},
