+++ inst/paper/paper.md
@@ -1,9 +1,11 @@
 ---
-title: 'nasapower: NASA-POWER Agroclimatology Data from R'
+title: 'nasapower: NASA-POWER Data from R'
 authors:
-  - affiliation: 1
-    name: Adam H Sparks
-    orcid: 0000-0002-0061-8359
+- affiliation: 1
+  name: Adam H Sparks
+  orcid: 0000-0002-0061-8359
+date: '2018-28-06'
+output: pdf_document
 bibliography: paper.bib
 tags:
 - NASA
@@ -11,6 +13,8 @@ tags:
 - meteorology
 - agroclimatology
 - climatology
+- alternative energy
+- sustainable buildings
 - R
 - climate data
 - earth science
@@ -18,27 +22,28 @@ tags:
 affiliations:
   index: 1
   name: University of Southern Queensland, Centre for Crop Health, Toowoomba Queensland 4350, Australia
-date: 2018-28-06
 ---
 
 # Summary and Statement of Need
 
-_nasapower_ [@sparks2018] is an R [@R-base] package providing functionality to
-interface with the NASA-POWER API [@stackhouseJr2018] for reproducible data
-retrieval using R. A single function, `get_power`, is provided. The function
-provides complete access to all functionality that the POWER API provides, which
-includes three user communities, AG (agroclimatoloy), SSE (Surface meteorology
-and Solar Energy) and SB (Sustainable Buildings); three temporal averages, Daily
+_nasapower_ is an R [@R-base] package providing functionality to interface with
+the NASA-POWER API [@stackhouseJr2018] for reproducible data retrieval using R.
+A single function, `get_power`, is provided. The function provides complete
+access to all functionality that the POWER API provides, which includes three
+user communities, AG (agroclimatoloy), SSE (Surface meteorology and Solar
+Energy) and SB (Sustainable Buildings); three temporal averages, Daily,
 Inter-annual and Climatology; three geographic options, single point, regional
 and global for the appropriate 141 parameters offered. Data are returned in a
-tidy data frame [@wickham2014] as a _tibble_ [@muller2018]. Integrating this
-data retrieval and formatting in R will streamline processes with models such as
-APSIM [@keating2003overview], DSSAT [@jones1998decision; @jones2003dssat] and
-EPIRICE [@savary2012] that can be linked to or are implemented fully in R.
-Extended documentation is provided with examples of converting it to spatial
-objects using _raster_ [@hijmans2017] and for obtaining solar radiation values
-and generating a .met file using the _APSIM_ R package [@fainges2017] to use in
-the APSIM model.
+tidy data frame [@wickham2014] as a _tibble_ [@muller2018]. Extended
+documentation is provided with examples of converting it to spatial objects
+using _raster_ [@hijmans2017] and for obtaining solar radiation values and
+generating a .met file using the _APSIM_ R package [@fainges2017] to use in the
+APSIM model.
+
+Integrating this data retrieval and formatting in R will streamline processes
+with models such as APSIM [@keating2003overview], DSSAT
+[@jones1998decision; @jones2003dssat] and EPIRICE [@savary2012] that can be
+linked to or are implemented fully in R.
 
 # About POWER Data
 
