### 2018-06-28 5a573d6
Remove references to "agroclimatology" only
- New package provides access to all three communities
