+++ README.Rmd
@@ -1,5 +1,5 @@
 ---
-title: "nasapower: NASA-POWER Agroclimatology Data from R"
+title: "nasapower: NASA-POWER Data from R"
 output: github_document
 ---
 
@@ -19,11 +19,11 @@ library(nasapower)
 
 <img align="right" src="man/figures/logo.png"> _nasapower_ aims to make
 it quick and easy to automate downloading
-[NASA-POWER](https://power.larc.nasa.gov) agroclimatology data in your R session
-as a tidy data frame for agricultural analysis and use in modelling or other
-purposes. POWER (Prediction Of Worldwide Energy Resource) data are freely
-available for download through a web interface at a resolution of 0.5 arc degree
-longitude by 0.5 arc degree latitude.
+[NASA-POWER](https://power.larc.nasa.gov) data in your R session as a tidy data
+frame for analysis and use in modelling or other purposes. POWER (Prediction Of
+Worldwide Energy Resource) data are freely available for download through a web
+interface at a resolution of 0.5 arc degree longitude by 0.5 arc degree
+latitude.
 
 Please see
 <https://power.larc.nasa.gov/> for more on the data and other ways to access it
