+++ README.Rmd
@@ -49,7 +49,7 @@ devtools::install_github("adamhsparks/nasapower", build_vignettes = TRUE)
 ## Introduction
 
 _nasapower_ aims to make it quick, easy and efficient to automate downloading
-NASA-POWER agroclimatology data in your R session as a tidy data frame.
+NASA-POWER data in your R session as a tidy data frame.
 
 _nasapower_ only provides one function, `get_power()`, which will download
 specified variables and return a tidy data frame of the requested data. Weather
