### 2018-06-28 6178237
Update formatting of README
