+++ README.Rmd
@@ -27,7 +27,7 @@ latitude.
 
 Please see
 <https://power.larc.nasa.gov/> for more on the data and other ways to access it
-and other forms of data available, e.g. an ESRI REST API.
+and other forms of data available, _e.g._ an ESRI REST API.
 
 ### Quick start
 
