+++ README.Rmd
@@ -130,4 +130,4 @@ gridded climate data sets that provide different data and different resolutions
 
 <https://power.larc.nasa.gov>
 
-<https://power.larc.nasa.gov/documents/Agroclimatology_Methodology.pdf>
+<https://power.larc.nasa.gov/documents/POWER_Data_v8_methodology.pdf>
