### 2018-06-28 #155 @adamhsparks
I've updated the initial on-boarding to reflect the newest version of the package now.
