### 2018-06-28 df187d2
Update description of NASA-POWER data
