+++ vignettes/nasapower.Rmd
@@ -23,7 +23,8 @@ knitr::opts_chunk$set(eval = FALSE,
 ## Introduction
 
 _nasapower_ aims to make it quick, easy and efficient to automate downloading
-NASA-POWER data in your R session as a tidy data frame.
+NASA-POWER global meteorology and surface solar energy data data in your R
+session as a tidy data frame.
 
 _nasapower_ only provides one function, `get_power()`, which will download
 specified variables and return a tidy data frame of the requested data. Weather
