+++ man/nasapower.Rd
@@ -6,14 +6,14 @@
 \alias{nasapower-package}
 \title{NASA-POWER Data from R}
 \description{
-Download NASA-POWER meteorology and climatology data for 141 parameters and
-create a tidy data frame for a single point, a region or global coverage.
-\pkg{nasapower} provides an R interface to the NASA - POWER API.  POWER
-(Prediction Of Worldwide Energy Resource) data are freely available for
-download with a resolution of 0.5 arc degree longitude by 0.5 arc degree
-latitude and are funded through the NASA Earth Science Directorate Applied
-Science Program. For more on the data themselves, please see
-\url{https://power.larc.nasa.gov/}.
+Download NASA-POWER global meteorology and surface solar energy climatology
+data for 141 parameters and create a tidy data frame for a single point, a
+region or global coverage. \pkg{nasapower} provides an R interface to the
+NASA - POWER API.  POWER (Prediction Of Worldwide Energy Resource) data are
+freely available for download with a resolution of 0.5 arc degree longitude
+by 0.5 arc degree latitude and are funded through the NASA Earth Science
+Directorate Applied Science Program. For more on the data themselves, please
+see \url{https://power.larc.nasa.gov/}.
 }
 \note{
 While \pkg{nasapower} does not redistribute the data in any way, we
