### 2018-06-28 7f98713
Fix missing "climatology" in description of data
