### 2018-06-28 3f5904d
Reformat note
