+++ man/nasapower.Rd
@@ -19,12 +19,11 @@ see \url{https://power.larc.nasa.gov/}.
 While \pkg{nasapower} does not redistribute the data in any way, we
 encourage users to follow the requests of the POWER Project Team.
 \preformatted{
-When POWER data products are used in a publication, we request the
-following acknowledgment be included:
+When POWER data products are used in a publication, we request the following
+acknowledgment be included:
 
-"These data were obtained from the NASA Langley Research Center
-POWER Project funded through the NASA Earth Science Directorate Applied
-Science Program."
+"These data were obtained from the NASA Langley Research Center POWER Project
+funded through the NASA Earth Science Directorate Applied Science Program."
 }
 }
 \references{
