### 2018-06-28 #155 @alisonboyer
Yes, I can review the new version.
--Alison
_________________________________
Alison Boyer
Research Staff, ORNL DAAC
Environmental Sciences Division
Oak Ridge National Laboratory

and

Joint Faculty Assistant Professor
Dept. of Ecology & Evolutionary Biology
University of Tennessee
Knoxville, Tennessee 37996

http://www.bio.utk.edu/boyer/index.html

On Mon, Jun 25, 2018 at 5:07 PM Scott Chamberlain <notifications@github.com>
wrote:

> Thanks @adamhsparks <https://github.com/adamhsparks> for the update.
>
> I dropped the status back to 3, reviewers assigned, I think we should get
> new reviews. Looks like @UniversalTourist
> <https://github.com/UniversalTourist> already said she can review again.
> @alisonboyer <https://github.com/alisonboyer> could you review this new
> version of the package? If not just let me know
>
