### 2018-07-05 c2fb3a1
Add checks for number of parameters to be downloaded
