### 2018-07-05 89a8cf1
Add checks for number of allowed `pars`
