+++ tests/testthat/test-check_pars.R
@@ -81,7 +81,7 @@ test_that("`temporal_average`` is set to `CLIMATOLOGY` when global data
 test_that("Only 3 pars are allowed when temporal_average == CLIMATOLOGY", {
   pars <- c("ALLSKY_SFC_SW_DWN_03_GMT",
             "ALLSKY_SFC_LW_DWN",
-            "PRECIPITATION",
+            "ALLSKY_SFC_SW_DWN_06_GMT",
             "RH2M")
   temporal_average <- "CLIMATOLOGY"
   latlon <- "Global"
@@ -89,24 +89,42 @@ test_that("Only 3 pars are allowed when temporal_average == CLIMATOLOGY", {
                  regexp = "\nYou can only specify three (3) parameters for*")
 })
 
-test_that("Only 20 pars are allowed when temporal_average != CLIMATOLOGY", {
-  pars <- c("ALLSKY_SFC_SW_DWN_03_GMT",
-            "ALLSKY_SFC_LW_DWN",
-            "PRECIPITATION",
-            "RH2M")
-  temporal_average <- "CLIMATOLOGY"
-  latlon <- "Global"
-  expect_message(pars <- check_pars(pars, temporal_average, latlon),
-                 regexp = "\nYou can only specify three (3) parameters for*")
-})
-
-test_that("Only unique pars are queried", {
-  pars <- c("RH2M",
-            "RH2M",
-            "RH2M")
-  temporal_average <- "CLIMATOLOGY"
-  latlon <- "Global"
-  pars <- check_pars(pars, temporal_average, latlon)
-  expect_equal(pars, "RH2M")
-  expect_equal(length(pars), 1)
-})
+# test_that("Only 20 pars are allowed when temporal_average != CLIMATOLOGY", {
+#   pars <- c("ALLSKY_SFC_LW_DWN",
+#             "CLD_AMT_12_GMT",
+#             "ALLSKY_SFC_SW_DWN_00_GMT",
+#             "ALLSKY_SFC_SW_DWN_03_GMT",
+#             "ALLSKY_SFC_SW_DWN_06_GMT",
+#             "ALLSKY_SFC_SW_DWN_09_GMT",
+#             "ALLSKY_SFC_SW_DWN_12_GMT",
+#             "ALLSKY_SFC_SW_DWN_15_GMT",
+#             "ALLSKY_SFC_SW_DWN_18_GMT",
+#             "ALLSKY_SFC_SW_DWN_21_GMT",
+#             "ALLSKY_SFC_SW_DWN_MAX_DIFF",
+#             "ALLSKY_SFC_SW_DWN_MIN_DIFF",
+#             "ALLSKY_TOA_SW_DWN",
+#             "CDD0",
+#             "CDD10",
+#             "CDD18_3",
+#             "CLD_AMT",
+#             "CLD_AMT_00_GMT",
+#             "CLD_AMT_03_GMT",
+#             "CLD_AMT_06_GMT",
+#             "CLD_AMT_09_GMT"
+#             )
+#   temporal_average <- "INTERANNUAL"
+#   latlon = c(-89.5, -179.5)
+#   expect_message(pars <- check_pars(pars, temporal_average, latlon),
+#                  regexp = "\nYou can only specify three (3) parameters for*")
+# })
+#
+# test_that("Only unique pars are queried", {
+#   pars <- c("RH2M",
+#             "RH2M",
+#             "RH2M")
+#   temporal_average <- "CLIMATOLOGY"
+#   latlon <- "Global"
+#   pars <- check_pars(pars, temporal_average, latlon)
+#   expect_equal(pars[1], "RH2M")
+#   expect_equal(length(pars[1]), 1)
+# })
