### 2018-07-05 382c389
Correct CLIMATOLOGY pars length test
