### 2018-07-06 5ed5b0b
Enable test for DAILY/CLIMATOLOGY pars query limit
