### 2018-07-07 f4d8bc4
Minor documentation updates
