+++ R/get_power.R
@@ -13,10 +13,11 @@
 #'  details for more.
 #' @param pars A character vector of solar, meteorological or climatology
 #'  parameters to download.  See \code{names(parameters)} for a full list of
-#'  valid values and definitions visit the
+#'  valid values and definitions.  Visit the
 #'  [POWER website](https://power.larc.nasa.gov/#resources) for the Parameter
-#'  Dictionary.  If downloading global coverage a maximum of 3 `pars` can be
-#'  specified at one time.
+#'  Dictionary.  If downloading CLIMATOLOGY a maximum of 3 `pars` can be
+#'  specified at one time for for DAILY and INTERANNUAL a maximum of 20 can
+#'  be specified at one time.
 #' @param dates A character vector of start and end dates in that order,\cr
 #'  *e.g.*, `dates = c("1983-01-01", "2017-12-31")`.  Not required for global
 #'  coverage.  See argument details for more.
