### 2018-07-19 8095981
Clean up README structure
