+++ README.md
@@ -1,41 +1,32 @@
-nasapower: NASA-POWER Data from R
-================
 
-[![Travis-CI Build
-Status](https://travis-ci.org/adamhsparks/nasapower.svg?branch=master)](https://travis-ci.org/adamhsparks/nasapower)
-[![AppVeyor Build
-Status](https://ci.appveyor.com/api/projects/status/github/adamhsparks/nasapower?branch=master&svg=true)](https://ci.appveyor.com/project/adamhsparks/nasapower)
-[![Coverage
-Status](https://img.shields.io/codecov/c/github/adamhsparks/nasapower/master.svg)](https://codecov.io/github/adamhsparks/nasapower?branch=master)
+[![Travis-CI Build Status](https://travis-ci.org/adamhsparks/nasapower.svg?branch=master)](https://travis-ci.org/adamhsparks/nasapower)
+[![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/adamhsparks/nasapower?branch=master&svg=true)](https://ci.appveyor.com/project/adamhsparks/nasapower)
+[![Coverage Status](https://img.shields.io/codecov/c/github/adamhsparks/nasapower/master.svg)](https://codecov.io/github/adamhsparks/nasapower?branch=master)
 [![DOI](https://zenodo.org/badge/109224461.svg)](https://zenodo.org/badge/latestdoi/109224461)
-[![Project Status: WIP – Initial development is in progress, but there
-has not yet been a stable, usable release suitable for the
-public.](http://www.repostatus.org/badges/latest/wip.svg)](http://www.repostatus.org/#wip)
+[![Project Status: WIP – Initial development is in progress, but there has not yet been a stable, usable release suitable for the public.](http://www.repostatus.org/badges/latest/wip.svg)](http://www.repostatus.org/#wip)
 
-## Introduction
 
-<img align="right" src="man/figures/logo.png"> *nasapower* aims to make
-it quick and easy to automate downloading
-[NASA-POWER](https://power.larc.nasa.gov) global meteorology and surface
-solar energy climatology data data in your R session as a tidy data
-frame for analysis and use in modelling or other purposes. POWER
-(Prediction Of Worldwide Energy Resource) data are freely available for
-download through a web interface at a resolution of 0.5 arc degree
-longitude by 0.5 arc degree latitude.
+# _nasapower_: NASA-POWER Data from R <img align="right" src="man/figures/logo.png">
 
-Please see <https://power.larc.nasa.gov/> for more on the data and other
-ways to access it and other forms of data available, *e.g.* an ESRI REST
-API.
+_nasapower_ aims to make it quick and easy to automate downloading
+[NASA-POWER](https://power.larc.nasa.gov) global meteorology and surface solar
+energy climatology data  data in your R session as a tidy data frame for
+analysis and use in modelling or other purposes. POWER (Prediction Of Worldwide
+Energy Resource) data are freely available for download through a web interface
+at a resolution of 0.5 arc degree longitude by 0.5 arc degree latitude. Other
+functionality is provided to quickly and easily generate weather.met files from
+the POWER data for use in the Agricultural Production Systems sIMulator (APSIM).
+
+Please see
+<https://power.larc.nasa.gov/> for more on the data and other ways to access it
+and other forms of data available, _e.g._ an ESRI REST API.
 
 ### Quick start
 
-With the changes to the NASA-POWER data and API, this package is a
-work-in-progress under active development, install at your own risk.
-
-*nasapower* is not available from CRAN, only GitHub. It can easily be
+_nasapower_ is not yet available from CRAN, only GitHub. It can easily be
 installed using the following code:
 
-``` r
+```r
[…]
