+++ README.Rmd
@@ -1,133 +1,0 @@
----
-title: "nasapower: NASA-POWER Data from R"
-output: github_document
----
-
-```{r setup, include=FALSE}
-knitr::opts_chunk$set(echo = TRUE)
-library(nasapower)
-```
-
-[![Travis-CI Build Status](https://travis-ci.org/adamhsparks/nasapower.svg?branch=master)](https://travis-ci.org/adamhsparks/nasapower)
-[![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/adamhsparks/nasapower?branch=master&svg=true)](https://ci.appveyor.com/project/adamhsparks/nasapower)
-[![Coverage Status](https://img.shields.io/codecov/c/github/adamhsparks/nasapower/master.svg)](https://codecov.io/github/adamhsparks/nasapower?branch=master)
-[![DOI](https://zenodo.org/badge/109224461.svg)](https://zenodo.org/badge/latestdoi/109224461)
-[![Project Status: WIP – Initial development is in progress, but there has not yet been a stable, usable release suitable for the public.](http://www.repostatus.org/badges/latest/wip.svg)](http://www.repostatus.org/#wip)
-
-
-## Introduction
-
-<img align="right" src="man/figures/logo.png"> _nasapower_ aims to make
-it quick and easy to automate downloading
-[NASA-POWER](https://power.larc.nasa.gov) global meteorology and surface solar
-energy climatology data  data in your R session as a tidy data frame for
-analysis and use in modelling or other purposes. POWER (Prediction Of Worldwide
-Energy Resource) data are freely available for download through a web interface
-at a resolution of 0.5 arc degree longitude by 0.5 arc degree latitude.
-
-Please see
-<https://power.larc.nasa.gov/> for more on the data and other ways to access it
-and other forms of data available, _e.g._ an ESRI REST API.
-
-### Quick start
-
-With the changes to the NASA-POWER data and API, this package is a
-work-in-progress under active development, install at your own risk.
-
-_nasapower_ is not available from CRAN, only GitHub. It can easily be installed
-using the following code:
-
-```{r, eval=FALSE}
-if (!require(devtools)) {
-  install.packages("devtools")
-  library(devtools)
-}
-
-devtools::install_github("adamhsparks/nasapower", build_vignettes = TRUE)
-```
-
-## Introduction
-
-_nasapower_ aims to make it quick, easy and efficient to automate downloading
-NASA-POWER data in your R session as a tidy data frame.
-
-_nasapower_ only provides one function, `get_power()`, which will download
-specified variables and return a tidy data frame of the requested data. Weather
-variables can be specified by using the `pars` argument. 
-
-## Using get_power()
-
-The `get_power()` function requires five arguments as seen in this example,
-which will fetch relative humidity (RH2M) and temperature (T2M) for the year
-1985 on a daily time-step.
-
-```{r example, eval=FALSE}
-power <- get_power(community = "AG",
-                   latlon = c(-89.5, -179.5),
-                   pars = c("RH2M", "T2M"),
-                   dates = c("1985-01-01", "1985-12-31"),
-                   temporal_average = "daily")
-```
-
-The arguments are:
-
-- `community`, a text string with valid values of: "AG" (Agroclimatology), "SSE"
-(Surface meteorology and Solar Energy) or "SB" (Sustainable Buildings). The
-selected user community will affect the units of the parameter and the temporal
-display of time series data (_e.g._ Agroclimatology will use MJ/m^2^/day for
-radiation units, while SSE and SB use kW/m^2^/day as units).
-
-- `latlon`, a length-2 numeric vector giving the decimal degree latitude and
-and longitude coordinates in that order for cell data to download or a length-4
[…]
