### 2018-07-19 c903121
Update documentation and internal object names
