+++ R/create_met.R
@@ -1,6 +1,4 @@
 
-
-
 #' @title Create an APSIM weather.met File from NASA - POWER Data
 #'
 #' @description Get NASA-POWER values for a single point or region and create
@@ -13,7 +11,15 @@
 #'  *e.g.*, `dates = c("1983-01-01", "2017-12-31")`.  See argument details for
 #'  more.
 #'
-#' @details Further details for each of the arguments are provided in their
+#' @details This function is essentially a wrapper for \code{\link{get_power}}
+#' and \code{\link[APSIM]{prepareMet}} that simplifies the querying of the
+#' POWER API and returns a .met object.
+#'
+#' The weather values from POWER for temperature are 2 metre max and min
+#' temperatures, T2M_MAX and T2M_MIN; radation, ALLSKY_SFC_SW_DWN; and rain,
+#' PRECTOT from the POWER AG community on a daily time-step.
+#'
+#' Further details for each of the arguments are provided in their
 #' respective sections following below.
 #'
 #' @section Argument details for `latlon`:
@@ -21,7 +27,7 @@
 #'  \item{For a single point}{To get a specific cell, 1/2 x 1/2 degree, supply a
 #'  length-2 numeric vector giving the decimal degree longitude and latitude in
 #'  that order for data to download,\cr
-#'  *e.g.*, `latlon = c(-179.5, -89.5)`.}
+#'  *e.g.*, `latlon = c(-27.48, 151.81)`.}
 #'
 #'  \item{For regional coverage}{To get a region, supply a length-4 numeric
 #'  vector as lower left (lat, lon) and upper right (lat, lon) coordinates,
@@ -38,15 +44,12 @@
 #'  If one date only is provided, it will be treated as both the start date and
 #'  the end date and only a single day's values will be returned.
 #'
-#' @references
-#' \url{https://power.larc.nasa.gov/documents/POWER_Data_v8_methodology.pdf}
-#'
 #' @examples
 #' # Create a .met object for Kingsthorpe, Qld from 1985-01-01 to 1985-06-30.
 #'
 #' \dontrun{
 #' Kingsthorpe <- create_met(latlon = c(-27.48, 151.81),
-#'                           dates = c("1985-01-01", "1985-06-30")
+#'                           dates = c("1985-01-01", "1985-12-31")
 #' )
 #' }
 #'
@@ -76,14 +79,14 @@ create_met <- function(latlon = NULL,
     )
   )
 
-  power_met <-
+  power_data <-
     dplyr::select(power_data,
-                  YEAR,
-                  DOY,
-                  T2M_MAX,
-                  T2M_MIN,
-                  PRECTOT,
-                  ALLSKY_SFC_SW_DWN)
+                  "YEAR",
+                  "DOY",
+                  "T2M_MAX",
+                  "T2M_MIN",
+                  "PRECTOT",
+                  "ALLSKY_SFC_SW_DWN")
 
   met_names <- c("year",
                  "day",
@@ -92,7 +95,7 @@ create_met <- function(latlon = NULL,
                  "radn",
                  "rain")
 
-  units <-
+  met_units <-
     c("()",
       "()",
       "(oC)",
@@ -100,13 +103,13 @@ create_met <- function(latlon = NULL,
       "(MJ/m^2/day)",
       "(mm)")
 
-  power_met <- APSIM::prepareMet(
-    power_met,
+  out <- APSIM::prepareMet(
+    power_data,
     lat = power_data[2, 1],
     lon = power_data[1, 1],
     newNames = met_names,
-    units = units
+    units = met_units
   )
 
-  return(power_met)
+  return(out)
 }
