### 2018-07-19 6b8c379
Fix in-function check for latlon
