### 2018-07-19 7950704
add tests for new function
