### 2018-07-19 98cef35
Update documentation for new functionality
