### 2018-07-19 6424166
Add link to APSIM.info from README
