### 2018-07-19 f7cd206
Refresh internal parameters file
