### 2018-07-19 e7edd94
Update paper with new functionality
