+++ README.md
@@ -10,13 +10,13 @@
 
 _nasapower_ aims to make it quick and easy to automate downloading
 [NASA-POWER](https://power.larc.nasa.gov) global meteorology and surface solar
-energy climatology data  data in your R session as a tidy data frame for
-analysis and use in modelling or other purposes. POWER (Prediction Of Worldwide
-Energy Resource) data are freely available for download through a web interface
-at a resolution of 0.5 arc degree longitude by 0.5 arc degree latitude. Other
-functionality is provided to quickly and easily generate a `metFile` object
-from POWER data for use in the [Agricultural Production Systems sIMulator
-(APSIM)](http://www.apsim.info/).
+energy climatology data  data in your R session as a tidy data frame `tibble`
+object for analysis and use in modelling or other purposes. POWER (Prediction Of
+Worldwide Energy Resource) data are freely available for download through a web
+interface at a resolution of 0.5 arc degree longitude by 0.5 arc degree
+latitude. Further functionality is provided to quickly and easily generate a
+`metFile` object from POWER data for use in the [Agricultural Production Systems
+sIMulator (APSIM)](http://www.apsim.info/).
 
 Please see
 <https://power.larc.nasa.gov/> for more on the data and other ways to access it
