### 2018-07-19 d0a09e2
Update vignette to follow a more standard format
