+++ vignettes/nasapower.Rmd
@@ -80,11 +80,11 @@ radiation units, while SSE and SB use kW/m^2^/day as units).
 - `temporal_average`, a character vector of the desired temporal average(s).
   Valid values are "DAILY", "INTERANNUAL" and "CLIMATOLOGY".
 
-## Creating .metFile Objects for Use with APSIM
+## Using create_met()
 
-Get NASA-POWER values for a single point, Kingsthorpe, Queensland for 1985 and
-create an APSIM weather.met object suitable for use in APSIM for crop modelling,
-saving it to disk using the _APSIM_ package.
+The `create_met()` function wraps `get_power()` and `prepareMet()` from the
+_APSIM_ package to simplify retrieving POWER data for use in the APSIM
+modelling framework.
 
 ```{r create_met, eval=FALSE}
 Kingsthorpe <- create_met(latlon = c(-27.48, 151.81),
@@ -94,6 +94,27 @@ Kingsthorpe <- create_met(latlon = c(-27.48, 151.81),
 library(APSIM)
 writeMetFile("Kingsthorpe.met", Kingsthorpe)
 ```
+
+Only two arguments are passed to this function, `latlon` and `dates`.
+
+- `latlon`
+  - *For a single point* To get a specific cell, 1/2 x 1/2 degree, supply a
+    length-2 numeric vector giving the decimal degree longitude and latitude in
+    that order for data to download,  *e.g.*, `latlon = c(-179.5, -89.5)`.  
+  
+  - *For regional coverage* To get a region, supply a length-4 numeric
+    vector as lower left (lat, lon) and upper right (lat, lon) coordinates,
+    *e.g.*, `latlon = c(ymin, xmin, ymax, xmax)` in that order for a given
+    region, *e.g.*, a bounding box for the south-western corner of Australia:
+    `latlon = c(-55.5, 112.5, -50.5, 115.5)`. *Max bounding box is 10 x 10
+    degrees* of 1/2 x 1/2 degree data, *i.e.*, 100 points maximum in total.  
+    
+    - `dates`, a vector of start and end dates for which to query the NASA-POWER
+  API, *e.g.*, `dates = c("1983-01-01", "2017-12-31")`. 
+
+Get NASA-POWER values for a single point, Kingsthorpe, Queensland for 1985 and
+create an APSIM weather.met object suitable for use in APSIM for crop modelling,
+saving it to disk using the _APSIM_ package.
 
 ## Creating Spatial Objects from `get_power`
 
