### 2018-07-20 fb0c9ca
try again to fix missing values on import
