### 2018-07-21 c29d252
swap lon/lat to follow normal software values
