+++ vignettes/use-case.Rmd
@@ -56,7 +56,7 @@ insolation incident on a horizontal surface, from NASA - POWER.
 
 ```{r get_data_show, eval=FALSE}
 sw_dwn <- get_power(
-  latlon = c(-27.48, 151.81),
+  lonlat = c(151.81, -27.48),
   dates = c("2010-01-01", "2010-12-31"),
   pars = "ALLSKY_SFC_SW_DWN",
   community = "AG",

[…]
