### 2018-07-21 58d7055
Fix bug in regional lonlat values for queries and tests
