### 2018-07-21 bfc6bb3
Update documentation for create_met
