### 2018-07-22 4aa201a
Fix bugs with lonlat testing
