### 2018-07-22 c438ead
Fix one more bug in tests
