### 2018-07-22 4bcfc4f
see if this fixes error related to message
