+++ NEWS.md
@@ -1,4 +1,4 @@
-# nasapower 1.0.0.9000
+# nasapower 1.0.0.9002
 
 ## Major changes
 
