### 2018-07-23 c7a347e
Update README
add link to APSIM R package in README
