### 2018-07-23 d0f438c
Update Vignette to match help files for consistency
