### 2018-07-23 476254d
Update help files to match Details section
Temporal average in CAPS
