+++ R/get_power.R
@@ -90,7 +90,7 @@
 #'                    lonlat = c(-179.5, -89.5),
 #'                    pars = c("RH2M", "T2M"),
 #'                    dates = "1985-01-01",
-#'                    temporal_average = "daily")
+#'                    temporal_average = "DAILY")
 #' }
 #'
 #' @author Adam H. Sparks, \email{adamhsparks@@gmail.com}
