### 2018-07-24 27e50d0
Add warning section to `tryCatch` for reading CSV file
Sometimes the server responds, but doesn't generate a CSV file. This warning should catch that and respond with an error message that a CSV file was not created.
