+++ R/internal_functions.R
@@ -396,10 +396,21 @@ power_query <- function(community,
                                 col_types = readr::cols(),
                                 na = "-99")
   }, # nocov start
+  # sometimes the server responds but doesn't provide a CSV file
+  warning = function(w) {
+    w$message <- paste(
+      "\nA CSV file was not created, this is a server error.",
+      "The server may not be responding or is currently responding improperly.",
+      "Please check <https://power.larc.nasa.gov/> for notifications if",
+      "you repeatedly get this error.",
+      sep = "\n"
+    )
+    stop(w)
+  },
   error = function(e) {
     # check if POWER is returning an error message
-    if (exists("txt") && "Alert" %in% names(txt)) {
-        e$message <- paste0(txt$messages$Alert$Description)
+    if ("messages" %in% names(txt)) {
+      e$message <- paste0(txt$messages$Alert$Description)
     } else {
       # if not, return our own error message
       e$message <- paste(
