### 2018-07-24 #155 @adamhsparks
Sorry, Scott, they can have more time.

NASA just made changes to the API suddenly without warning. The formerly proper CSV file now is no longer rectangular, I need to rethink how the data is ingested and I'm leaving for two weeks of travel on Friday. I've got some ideas in mind but it'll take me a while before I can get to it.
