### 2018-07-25 fd5f51f
Update README.md
Add statement about the package not working currently
