### 2018-07-27 6fcfcfd
Fix issues reading csv from server
this should automatically skip the new header file. However, the server won't respond now for me to check
