### 2018-08-01 #155 @adamhsparks
Sorry about that. I think we're ready now again for review. The POWER team changed the response of the API to include some extra metadata and didn't alert anyone interfacing with the API. I've now made the changes to be able to gather that metadata but also fixed the issue that precluded the data from being read due to changes in the response.
