### 2018-08-12 #155 @UniversalTourist
I finally have time today to finish the review! I just wanted to let you know about that :)
Many things has changed apparently! :)
