### 2018-08-12 #155 @UniversalTourist
Estimated hours spent reviewing: It took almost, totally, 2 hours to check everything.
---

- Installation:

With MacOS:
I used macOS High Sierra, Verison 10.13.3 and R version 3.5.0 (Joy in Playing). I was able to install the package from Github without any problem by devtools.

- Code Overall:
These package has one function to get data for a cell or for a region. It has nice structered, easy to follow and also nice explained with comments. Also two new functions which help using nasa power data for modelling purpose. I don't have any experience on ICASA file for use in DSSAT modelling but I used APSIM files for an archaelogical modelling project two years ago. And I found it very useful.

- rOpenSci criteria:
Package name is really nice. Function and variable naming is proper. Licence and Code of Conduct are available.

- Readme:
Readme.rmd is clear and easy to follow steps for installations.

- Errors and console messages:
They are all clear.

- R CMD check:

I only get these error and warning. I always see that error when I'm running R but I don't know the reason for warning.

```
Status: 1 ERROR, 1 WARNING
See
  '/private/var/folders/9g/cqymx4b10nd1_45vjh6zq_040000gn/T/RtmpqYWghO/nasapower.Rcheck/00check.log'

R CMD check results
0 errors | 1 warning  | 0 notes
checking DESCRIPTION meta-information ... WARNING
During startup - Warning messages:
1: Setting LC_CTYPE failed, using "C"
2: Setting LC_TIME failed, using "C"
3: Setting LC_MESSAGES failed, using "C"
4: Setting LC_MONETARY failed, using "C"

ERROR
During startup - Warning messages:
1: Setting LC_CTYPE failed, using "C"
2: Setting LC_TIME failed, using "C"
3: Setting LC_MESSAGES failed, using "C"
4: Setting LC_MONETARY failed, using "C"
```
