### 2018-08-12 #155 @adamhsparks
Hi @UniversalTourist, try visiting this SO thread to see if it helps?

https://stackoverflow.com/questions/9689104/installing-r-on-mac-warning-messages-setting-lc-ctype-failed-using-c

As you've indicated, it always happens, so it shouldn't be related to `nasapower`. :)
