### 2018-09-19 #155 @sckott
Good idea to add guidelines for users on how to contribute. See e.g. https://github.com/ropensci/dotgithubfiles/

---

awesomeness:
- nice work!
- nice failure behavior!

other comments:
- duplicate `data` in REAMDE _surface solar energy climatology data  data_
- you could be a bit more detailed on your `parameters` man file, e.g., `A list with 141 objects`. What kind of objects?  And it looks like the list is named by something useful?
- Having examples for all functions is good, but the examples are pretty sparse, only 1 for each fxn.
- including `@return` for each fxn would help users know what to expect
- nice use of parameter checking in your fxns. for logical parameters, there doesn't seem to be checks. so e.g., a user could do `get_power(meta = 5)`, though unlikely could happen i guess
- it's probably a good idea for examples and tests to write to tmp dir instead of a non-temporary user path, e.g., `dsn = "~/Documents"`
- this failure behavior isn't great when `dsn` and or `file_out` not given (same for `create_met`):

```r
create_icasa(lonlat = c(151.81, -27.48),
             dates = c("1985-01-01", "1985-12-31")
)
Error in if (substr(file_out, nchar(file_out) - 3, nchar(file_out)) !=  :
  argument is of length zero
```
