### 2018-09-19 #155 @sckott
I'm sure I have some fxns that return different objects depending on user inputs, but it definitely seems better to avoid that. Maybe one of these two optoins:

- put the metadata in attributes in the data.frame, then user can access them with `attributes` or some custom thing
- make a custom print fxn where the metadata is printed on top of the data.frame sort of like tibbles do, then the object itself is still a data.frame
