### 2018-09-19 #155 @adamhsparks
I've some ideas of what can be done with it. I'll think on it some more and see what I can implement. I think I've addressed everything else.
