### 2018-09-20 #155 @adamhsparks
No, please wait, I've pushed a custom `print()` method last night that includes an attribute header to the `devel` branch. Polishing it, will be done within 48 hrs I'd guess?
