### 2018-09-21 #155 @sckott
@adamhsparks a few things:

- the example for `create_icasa` has `file = ICASA_example.txt` where the file name is not quoted, R is looking for an object named `ICASA_example.txt`
- for the `create_met` example you have `dsn = tmpdir()`, which sholuld be `tempdir()` -
- the `get_power` 2nd example doesn't work either. you have `temporal_average` as a required parameter, but don't supply anything to it in the 2nd example.
- Also about `get_power`. it's not a good idea to have a required parameter (w/o a default) after a non-required parameter. that is, you have `function(community, lonlat, pars, dates = NULL, temporal_average)`, whereas `function(community, lonlat, pars, temporal_average, dates = NULL)` would make more sense b/c all required paramters are first, then the last is optional

Its essential that your examples run successfully. a good way to do this is have them run on travis, e.g. https://github.com/ropensci/wellknown/blob/master/.travis.yml#L28 Please make sure your examples work moving forward
