### 2018-09-21 #155 @adamhsparks
Thanks @sckott, I'll address these issues.
