### 2018-09-24 #155 @sckott
nice, glad it was helpful.

having another quick look
