### 2018-09-24 #155 @sckott
Approved!  Thanks again for your submission @adamhsparks  !

* Please transfer the package  to `ropensci`- You'll be made admin once you do.
* Make sure to
    * add rOpenSci footer to README
`[![ropensci_footer](https://ropensci.org/public_images/ropensci_footer.png)](https://ropensci.org)`

---

* Are you interested in doing a blog post for our blog https://ropensci.org/blog/ ?  either a short-form intro to it (https://ropensci.org/technotes/) or long-form post with more narrative about its development (https://ropensci.org/blog/). If so, we'll have @stefaniebutland get in touch with you
